% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{check_value}
\alias{check_value}
\title{check value}
\usage{
check_value(input)
}
\arguments{
\item{input}{- DeepBlue XML-RPC return to be checked.}
}
\description{
return the function result and check for error message from DeepBlue server.
}

