% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{createBody}
\alias{createBody}
\title{createBody}
\usage{
createBody(method, args)
}
\arguments{
\item{method}{DeepBlue operation to be executed}

\item{args}{Operation arguments}
}
\description{
Create XML-RPC request body document
}

