% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue.add_annotation}
\alias{deepblue.add_annotation}
\title{add_annotation}
\usage{
deepblue.add_annotation(name = NULL, genome = NULL, description = NULL,
  data = NULL, format = NULL, extra_metadata = NULL,
  user_key = deepblue.USER_KEY)
}
\arguments{
\item{name}{- A string (annotation name)}

\item{genome}{- A string (the target genome)}

\item{description}{- A string (description of the annotation)}

\item{data}{- A string (the BED formatted data)}

\item{format}{- A string (format of the provided data)}

\item{extra_metadata}{- A struct (additional metadata)}

\item{user_key}{- A string (users token key)}
}
\value{
id - A string (id of the newly inserted annotation)
}
\description{
Inserts a new annotation with the given parameters.
}
\examples{
\dontrun{
  data = "chr1	28735	29810
chr1	135124	135563
chr1	327790	328229
chr1	437151	438164
chr1	449273	450544
chr1	533219	534114
chr1	544738	546649
chr1	713984	714547
chr1	762416	763445
chr1	788863	789211"
deepblue.add_annotation(name="Interesting annotation", genome="hg19",
	description="It is an annotation with some interesting regions",
	data=data, format="CHROMOSOME,START,END",
	extra_metadata=list(source="My own experiments"),
	user_key="my_private_user_key")
}

}
\seealso{
Other Inserting and listing annotations: \code{\link{deepblue.find_pattern}},
  \code{\link{deepblue.list_annotations}}
}

