% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.r
\name{deepblue.add_biosource}
\alias{deepblue.add_biosource}
\title{add_biosource}
\usage{
deepblue.add_biosource(name = NULL, description = NULL,
  extra_metadata = NULL, user_key = deepblue.USER_KEY)
}
\arguments{
\item{name}{- A string (biosource name)}

\item{description}{- A string (description of the biosource)}

\item{extra_metadata}{- A struct (additional metadata)}

\item{user_key}{- A string (users token key)}
}
\value{
a vector with  ('okay',
}
\description{
Inserts a new biosource with the given parameters.

id of the newly inserted biosource) or ('error', error_message)
}
\seealso{
Other Inserting and listing biosources: \code{\link{deepblue.list_biosources}},
  \code{\link{deepblue.list_similar_biosources}}
}

