% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue.add_epigenetic_mark}
\alias{deepblue.add_epigenetic_mark}
\title{add_epigenetic_mark}
\usage{
deepblue.add_epigenetic_mark(name = NULL, description = NULL,
  extra_metadata = NULL, user_key = deepblue.USER_KEY)
}
\arguments{
\item{name}{- A string (name of the epigenetic mark)}

\item{description}{- A string (description of the epigenetic mark)}

\item{extra_metadata}{- A struct (additional metadata)}

\item{user_key}{- A string (users token key)}
}
\value{
id - A string (id of the newly inserted epigenetic mark)
}
\description{
Inserts a new epigenetic mark with the given parameters.
}
\examples{
deepblue.add_epigenetic_mark(name="H3k123ac Name",
	description="A brand new histone mark",
	extra_metadata=list(type="Histone Mark"),
	user_key="my_secret_key" )

}
\seealso{
Other Inserting and listing epigenetic marks: \code{\link{deepblue.list_epigenetic_marks}},
  \code{\link{deepblue.list_similar_epigenetic_marks}}
}

