% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue.add_experiment}
\alias{deepblue.add_experiment}
\title{add_experiment}
\usage{
deepblue.add_experiment(name = NULL, genome = NULL,
  epigenetic_mark = NULL, sample = NULL, technique = NULL,
  project = NULL, description = NULL, data = NULL, format = NULL,
  extra_metadata = NULL, user_key = deepblue.USER_KEY)
}
\arguments{
\item{name}{- A string (experiment name)}

\item{genome}{- A string (the target genome)}

\item{epigenetic_mark}{- A string (epigenetic mark of the experiment)}

\item{sample}{- A string (id of the used sample)}

\item{technique}{- A string (technique used by this experiment)}

\item{project}{- A string (the project name)}

\item{description}{- A string (description of the experiment)}

\item{data}{- A string (the BED formated data)}

\item{format}{- A string (format of the provided data)}

\item{extra_metadata}{- A struct (additional metadata)}

\item{user_key}{- A string (users token key)}
}
\value{
id - A string (id of the newly inserted experiment)
}
\description{
Inserts a new experiment with the given parameters.
}
\examples{
\dontrun{
  data = "chr1	28735	29810
chr1	135124	135563
chr1	327790	328229
chr1	437151	438164
chr1	449273	450544
chr1	533219	534114
chr1	544738	546649
chr1	713984	714547
chr1	762416	763445
chr1	788863	789211"
deepblue.add_experiment(name="Interesting experiment", genome="hg19",
	epigenetic_mark="H3k27ac", sample="s123456", technique="Chip-seq",
	project="My Own project",
	description="It is an experiment with some interesting regions", data=data,
	format="CHROMOSOME,START,END",
	extra_metadata=list(source="My own experiments"),
	user_key="my_private_user_key")
}

}
\seealso{
Other Inserting and listing experiments: \code{\link{deepblue.collection_experiments_count}},
  \code{\link{deepblue.faceting_experiments}},
  \code{\link{deepblue.list_experiments}},
  \code{\link{deepblue.list_recent_experiments}},
  \code{\link{deepblue.list_similar_experiments}}
}

