% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue.add_gene_set}
\alias{deepblue.add_gene_set}
\title{add_gene_set}
\usage{
deepblue.add_gene_set(name = NULL, description = NULL, data = NULL,
  format = NULL, extra_metadata = NULL, user_key = deepblue.USER_KEY)
}
\arguments{
\item{name}{- A string (gene set name)}

\item{description}{- A string (description of the annotation)}

\item{data}{- A string (the BED formatted data)}

\item{format}{- A string (Currently, it is only supported GTF.)}

\item{extra_metadata}{- A struct (additional metadata)}

\item{user_key}{- A string (users token key)}
}
\value{
id - A string (id of the newly inserted annotation)
}
\description{
Inserts a new set of genes in the GTF format. Important: It will only include the rows that have 'gene' as feature.
}
\seealso{
Other Operations on gene sets and genes identifiers: \code{\link{deepblue.select_genes}}
}

