% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue.add_project}
\alias{deepblue.add_project}
\title{add_project}
\usage{
deepblue.add_project(name = NULL, description = NULL,
  user_key = deepblue.USER_KEY)
}
\arguments{
\item{name}{- A string (projectname)}

\item{description}{- A string (description of the project)}

\item{user_key}{- A string (users token key)}
}
\value{
id - A string (id of the newly inserted project)
}
\description{
Inserts a new project with the given parameters.
}
\seealso{
Other Inserting and listing projects: \code{\link{deepblue.add_user_to_project}},
  \code{\link{deepblue.list_projects}},
  \code{\link{deepblue.list_similar_projects}},
  \code{\link{deepblue.set_project_public}}
}

