% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.r
\name{deepblue.add_sample}
\alias{deepblue.add_sample}
\title{add_sample}
\usage{
deepblue.add_sample(biosource_name = NULL, extra_metadata = NULL,
  user_key = deepblue.USER_KEY)
}
\arguments{
\item{biosource_name}{- A string (biosource name)}

\item{extra_metadata}{- A struct (sample extra metadata. You can include any key-value collection here.)}

\item{user_key}{- A string (users token key)}
}
\value{
a vector with  ('okay',
}
\description{
Inserts a new sample of a given biosourcea.

id of the newly inserted sample) or ('error', error_message)
}
\seealso{
Other Inserting and listing samples: \code{\link{deepblue.add_sample_from_gsm}},
  \code{\link{deepblue.list_samples}}
}

