% Generated by roxygen2: do not edit by hand
% Please edit documentation in inst/deepblue.R
\name{deepblue.add_user_to_project}
\alias{deepblue.add_user_to_project}
\title{add_user_to_project}
\usage{
deepblue.add_user_to_project(user = NULL, project = NULL, set = NULL,
  user_key = deepblue.USER_KEY)
}
\arguments{
\item{user}{- A string (User name or ID)}

\item{project}{- A string (Project name or ID)}

\item{set}{- A boolean (True to include the user or false to remove)}

\item{user_key}{- A string (users token key)}
}
\value{
user_id - A string (id of the user)
}
\description{
Include or exclude an user from a project
}
\seealso{
Other Inserting and listing projects: \code{\link{deepblue.add_project}},
  \code{\link{deepblue.list_projects}},
  \code{\link{deepblue.list_similar_projects}},
  \code{\link{deepblue.set_project_public}}
}

