% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue.aggregate}
\alias{deepblue.aggregate}
\title{aggregate}
\usage{
deepblue.aggregate(data_id = NULL, ranges_id = NULL, column = NULL,
  user_key = deepblue.USER_KEY)
}
\arguments{
\item{data_id}{- A string (id of the query with the data)}

\item{ranges_id}{- A string (id of the query with the regions range)}

\item{column}{- A string (name of the column that will be used in the aggregation)}

\item{user_key}{- A string (users token key)}
}
\value{
regions - A string (query id of this aggregation operation)
}
\description{
Summarize the data regions content in range regions. Use the fields @AGG.MIN, @AGG.MAX, @AGG.MEDIAN, @AGG.MEAN, @AGG.VAR, @AGG.SD, @AGG.COUNT in the get_regions command format parameter for retrieving the computed values.
}
\examples{
annotation_id = deepblue.select_annotations(annotation_name="CpG Islands",
	genome="hg19", chromosome="chr1")
data_id = deepblue.select_experiments(
	experiment_name="E002-H3K9ac.narrowPeak.bed")
deepblue.aggregate(data_id = data_id, ranges_id=annotation_id,
	column = "SCORE")

}
\seealso{
Other Operating on the data regions: \code{\link{deepblue.count_regions}},
  \code{\link{deepblue.extend}},
  \code{\link{deepblue.filter_regions}},
  \code{\link{deepblue.flank}},
  \code{\link{deepblue.get_experiments_by_query}},
  \code{\link{deepblue.get_regions}},
  \code{\link{deepblue.input_regions}},
  \code{\link{deepblue.intersection}},
  \code{\link{deepblue.merge_queries}},
  \code{\link{deepblue.query_cache}},
  \code{\link{deepblue.query_experiment_type}},
  \code{\link{deepblue.score_matrix}},
  \code{\link{deepblue.select_annotations}},
  \code{\link{deepblue.select_experiments}},
  \code{\link{deepblue.select_regions}},
  \code{\link{deepblue.tiling_regions}}
}

