% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue.cancel_request}
\alias{deepblue.cancel_request}
\title{cancel_request}
\usage{
deepblue.cancel_request(id = NULL, user_key = deepblue.USER_KEY)
}
\arguments{
\item{id}{- A string (Request ID to be canceled, stopped or removed.)}

\item{user_key}{- A string (users token key)}
}
\value{
id - A string (ID of the canceled request)
}
\description{
Stop, cancel, and remove request data. Its data will be remove if the request did finish.
}
\examples{
deepblue.cancel_request(id = "r12345", user_key = "my_secret_key")

}
\seealso{
Other Commands for all types of data: \code{\link{deepblue.info}},
  \code{\link{deepblue.is_biosource}},
  \code{\link{deepblue.list_in_use}},
  \code{\link{deepblue.remove}},
  \code{\link{deepblue.search}}
}

