% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue.chromosomes}
\alias{deepblue.chromosomes}
\title{chromosomes}
\usage{
deepblue.chromosomes(genome = NULL, user_key = deepblue.USER_KEY)
}
\arguments{
\item{genome}{- A string (the target genome)}

\item{user_key}{- A string (users token key)}
}
\value{
chromosomes - A array (A list containing all chromosomes, with theirs names and sizes)
}
\description{
List all chromosomes of a given genome.
}
\examples{
\dontrun{
  deepblue.chromosomes(genome = "g1", user_key = deepblue.USER_KEY)
}

}
\seealso{
Other Inserting and listing genomes: \code{\link{deepblue.add_genome}},
  \code{\link{deepblue.list_genomes}},
  \code{\link{deepblue.list_similar_genomes}},
  \code{\link{deepblue.upload_chromosome}}
}

