% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue.clone_dataset}
\alias{deepblue.clone_dataset}
\title{clone_dataset}
\usage{
deepblue.clone_dataset(dataset_id = NULL, new_name = NULL,
  new_epigenetic_mark = NULL, new_sample = NULL, new_technique = NULL,
  new_project = NULL, description = NULL, format = NULL,
  extra_metadata = NULL, user_key = deepblue.USER_KEY)
}
\arguments{
\item{dataset_id}{- A string (ID of the dataset (experiment or annotation ID))}

\item{new_name}{- A string (New dataset name)}

\item{new_epigenetic_mark}{- A string (New epigenetic mark)}

\item{new_sample}{- A string (New sample ID)}

\item{new_technique}{- A string (New technique)}

\item{new_project}{- A string (New project)}

\item{description}{- A string (description of the experiment - empty to copy from the cloned dataset)}

\item{format}{- A string (format of the provided data - empty to copy from the cloned dataset)}

\item{extra_metadata}{- A struct (additional metadata - empty to copy from the cloned dataset)}

\item{user_key}{- A string (users token key)}
}
\value{
id - A string (id of the new dataset)
}
\description{
Clone the dataset, allowing to change the description, column format (restrictively), and extra_metadata.
}
\seealso{
Other Operations that modify the data content: \code{\link{deepblue.change_extra_metadata}}
}

