% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue.commands}
\alias{deepblue.commands}
\title{commands}
\usage{
deepblue.commands()
}
\value{
commands - A struct (command descriptions)
}
\description{
Lists all existing commands.
}
\examples{
deepblue.commands()

}
\seealso{
Other Checking DeepBlue status: \code{\link{deepblue.echo}}
}

