% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalFunctions.R
\name{deepblue.convert_to_df}
\alias{deepblue.convert_to_df}
\title{convert_to_df}
\usage{
deepblue.convert_to_df(output, inf, dict = col_dict)
}
\arguments{
\item{output}{A string}

\item{inf}{A list with request information}
}
\value{
regions A data frame
}
\description{
save output in a data frame for further processing.Expects two parameters; the output string from method deepblue.get_request_data and request information from method process_request.
}

