% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue.count_regions}
\alias{deepblue.count_regions}
\title{count_regions}
\usage{
deepblue.count_regions(query_id = NULL, user_key = deepblue.USER_KEY)
}
\arguments{
\item{query_id}{- A string (Query ID)}

\item{user_key}{- A string (users token key)}
}
\value{
request_id - A string (Request ID - Use it to retrieve the result with info() and get_request_data())
}
\description{
Send a request to count the number of regions in the result of the given query.
}
\examples{
\dontrun{
  deepblue.count_regions(query_id = "q12345", user_key = deepblue.USER_KEY)
}

}
\seealso{
Other Operating on the data regions: \code{\link{deepblue.aggregate}},
  \code{\link{deepblue.extend}},
  \code{\link{deepblue.filter_regions}},
  \code{\link{deepblue.flank}},
  \code{\link{deepblue.get_experiments_by_query}},
  \code{\link{deepblue.get_regions}},
  \code{\link{deepblue.input_regions}},
  \code{\link{deepblue.intersection}},
  \code{\link{deepblue.merge_queries}},
  \code{\link{deepblue.query_cache}},
  \code{\link{deepblue.query_experiment_type}},
  \code{\link{deepblue.score_matrix}},
  \code{\link{deepblue.select_annotations}},
  \code{\link{deepblue.select_experiments}},
  \code{\link{deepblue.select_regions}},
  \code{\link{deepblue.tiling_regions}}
}

