% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue.create_column_type_calculated}
\alias{deepblue.create_column_type_calculated}
\title{create_column_type_calculated}
\usage{
deepblue.create_column_type_calculated(name = NULL, description = NULL,
  code = NULL, user_key = deepblue.USER_KEY)
}
\arguments{
\item{name}{- A string (column type name)}

\item{description}{- A string (description of the column type)}

\item{code}{- A string (Lua code that will be executed)}

\item{user_key}{- A string (users token key)}
}
\value{
id - A string (id of the newly created column type)
}
\description{
Create a calculated column
}
\examples{
deepblue.create_column_type_calculated(name = "CALC_SQUARE_ROOT_VALUE",
  description ="The square root of the value column",
  code = "return sqrt(value_of('VALUE'))", user_key = "my_secret_key")

}
\seealso{
Other Inserting and listing different column types: \code{\link{deepblue.create_column_type_category}},
  \code{\link{deepblue.create_column_type_range}},
  \code{\link{deepblue.create_column_type_simple}},
  \code{\link{deepblue.list_column_types}}
}

