% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue.create_column_type_range}
\alias{deepblue.create_column_type_range}
\title{create_column_type_range}
\usage{
deepblue.create_column_type_range(name = NULL, description = NULL,
  minimum = NULL, maximum = NULL, user_key = deepblue.USER_KEY)
}
\arguments{
\item{name}{- A string (column type name)}

\item{description}{- A string (description of the column type)}

\item{minimum}{- A double (minimum value for this range (inclusive))}

\item{maximum}{- A double (maximum value for this range (inclusive))}

\item{user_key}{- A string (users token key)}
}
\value{
id - A string (id of the newly created column type)
}
\description{
Create a column type from a category set.
}
\seealso{
Other Inserting and listing different column types: \code{\link{deepblue.create_column_type_calculated}},
  \code{\link{deepblue.create_column_type_category}},
  \code{\link{deepblue.create_column_type_simple}},
  \code{\link{deepblue.list_column_types}}
}

