% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue.create_column_type_simple}
\alias{deepblue.create_column_type_simple}
\title{create_column_type_simple}
\usage{
deepblue.create_column_type_simple(name = NULL, description = NULL,
  type = NULL, user_key = deepblue.USER_KEY)
}
\arguments{
\item{name}{- A string (column type name)}

\item{description}{- A string (description of the column type)}

\item{type}{- A string (type of the column type (string, integer, double))}

\item{user_key}{- A string (users token key)}
}
\value{
id - A string (id of the newly created column type)
}
\description{
Create a column type from a category set.
}
\examples{
deepblue.create_column_type_simple(name = "SOME_VALUE",
	description = "This column will include some value",
  	type = "integer", user_key = "my_secret_key")

}
\seealso{
Other Inserting and listing different column types: \code{\link{deepblue.create_column_type_calculated}},
  \code{\link{deepblue.create_column_type_category}},
  \code{\link{deepblue.create_column_type_range}},
  \code{\link{deepblue.list_column_types}}
}

