% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\docType{data}
\name{deepblue.debug.VERBOSE}
\alias{deepblue.debug.VERBOSE}
\title{Verbose}
\format{An object of class \code{logical} of length 1.}
\usage{
deepblue.debug.VERBOSE
}
\description{
Show or hide debugging messages
}
\keyword{datasets}

