% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue.echo}
\alias{deepblue.echo}
\title{echo}
\usage{
deepblue.echo(user_key = deepblue.USER_KEY)
}
\arguments{
\item{user_key}{- A string (users token key)}
}
\value{
message - A string (echo message including version)
}
\description{
Echos the server's version.
}
\seealso{
Other Checking DeepBlue status: \code{\link{deepblue.commands}}
}

