% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue.extend}
\alias{deepblue.extend}
\title{extend}
\usage{
deepblue.extend(query_id = NULL, length = NULL, direction = NULL,
  use_strand = NULL, user_key = deepblue.USER_KEY)
}
\arguments{
\item{query_id}{- A string (id of the query that contains the regions)}

\item{length}{- A int (The new region length)}

\item{direction}{- A string (The direction that the region will be extended: 'BACKWARD', 'FORWARD', 'BOTH'. (Empty value will be used for both direction.)}

\item{use_strand}{- A boolean (Use the region column STRAND to define the region direction)}

\item{user_key}{- A string (users token key)}
}
\value{
id - A string (id of the new query)
}
\description{
Extend regions.
}
\examples{
\dontrun{
  deepblue.extend(query_id = "q1234", length = 2000, direction = "BOTH",
  use_strand = TRUE, user_key = "my_secret_key")
}

}
\seealso{
Other Operating on the data regions: \code{\link{deepblue.aggregate}},
  \code{\link{deepblue.count_regions}},
  \code{\link{deepblue.filter_regions}},
  \code{\link{deepblue.flank}},
  \code{\link{deepblue.get_experiments_by_query}},
  \code{\link{deepblue.get_regions}},
  \code{\link{deepblue.input_regions}},
  \code{\link{deepblue.intersection}},
  \code{\link{deepblue.merge_queries}},
  \code{\link{deepblue.query_cache}},
  \code{\link{deepblue.query_experiment_type}},
  \code{\link{deepblue.score_matrix}},
  \code{\link{deepblue.select_annotations}},
  \code{\link{deepblue.select_experiments}},
  \code{\link{deepblue.select_regions}},
  \code{\link{deepblue.tiling_regions}}
}

