% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue.extract_names}
\alias{deepblue.extract_names}
\title{extract_names}
\usage{
deepblue.extract_names(list = NULL)
}
\arguments{
\item{list}{- A array (list of lists of IDs and Names)}
}
\value{
names - A array (list containing the extracted names)
}
\description{
Extract the names from a list of ID and Names.
}
\examples{
deepblue.extract_names(list = list(list("a124", "Annotation 1"), list("a1235", "Annotation 2")))

}
\seealso{
Other Utilities for connecting operations: \code{\link{deepblue.extract_ids}}
}

