% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue.filter_regions}
\alias{deepblue.filter_regions}
\title{filter_regions}
\usage{
deepblue.filter_regions(query_id = NULL, field = NULL, operation = NULL,
  value = NULL, type = NULL, user_key = deepblue.USER_KEY)
}
\arguments{
\item{query_id}{- A string (id of the query to be filtered)}

\item{field}{- A string (field that is filtered by)}

\item{operation}{- A string (operation used for filtering. For 'string' must be '==' or '!=' and for 'number' must be one of these: ==,!=,>,>=,<,<=)}

\item{value}{- A string (value the operator is applied to)}

\item{type}{- A string (type of the value: 'number' or 'string' )}

\item{user_key}{- A string (users token key)}
}
\value{
id - A string (id of filtered query)
}
\description{
Filters the result of the given query by the given restrictions.
}
\seealso{
Other Operating on the data regions: \code{\link{deepblue.aggregate}},
  \code{\link{deepblue.count_regions}},
  \code{\link{deepblue.extend}},
  \code{\link{deepblue.flank}},
  \code{\link{deepblue.get_experiments_by_query}},
  \code{\link{deepblue.get_regions}},
  \code{\link{deepblue.input_regions}},
  \code{\link{deepblue.intersection}},
  \code{\link{deepblue.merge_queries}},
  \code{\link{deepblue.query_cache}},
  \code{\link{deepblue.query_experiment_type}},
  \code{\link{deepblue.score_matrix}},
  \code{\link{deepblue.select_annotations}},
  \code{\link{deepblue.select_experiments}},
  \code{\link{deepblue.select_regions}},
  \code{\link{deepblue.tiling_regions}}
}

