% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue.find_pattern}
\alias{deepblue.find_pattern}
\title{find_pattern}
\usage{
deepblue.find_pattern(pattern = NULL, genome = NULL, overlap = NULL,
  user_key = deepblue.USER_KEY)
}
\arguments{
\item{pattern}{- A string (pattern (regular expression))}

\item{genome}{- A string (the target genome)}

\item{overlap}{- A boolean (if the matching should do overlap search)}

\item{user_key}{- A string (users token key)}
}
\value{
id - A string (id of the annotation that contains the positions of the given pattern)
}
\description{
Process an annotation that will contain all genomic positions from the given pattern.
}
\examples{
\dontrun{
  deepblue.find_pattern(pattern = "CAC[TAG]+CAC", genome = "hg19", overlap = FALSE,
  user_key = "my_secret_key")
}

}
\seealso{
Other Inserting and listing annotations: \code{\link{deepblue.add_annotation}},
  \code{\link{deepblue.list_annotations}}
}

