% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue.flank}
\alias{deepblue.flank}
\title{flank}
\usage{
deepblue.flank(query_id = NULL, start = NULL, length = NULL,
  use_strand = NULL, user_key = deepblue.USER_KEY)
}
\arguments{
\item{query_id}{- A string (id of the query that contains the regions)}

\item{start}{- A int (Number of base pairs after the end of the region. Use a negative number to denote the number of base pairs before the start of the region.)}

\item{length}{- A int (The new region length)}

\item{use_strand}{- A boolean (Use the region column STRAND to define the region direction)}

\item{user_key}{- A string (users token key)}
}
\value{
id - A string (id of the new query)
}
\description{
Generate flanking regions from the given regions.
}
\examples{
annotation_id = deepblue.select_annotations(annotation_name="CpG Islands",
	genome="hg19", chromosome="chr1")
deepblue.flank(query_id = annotation_id, start = 0, length = 2000,
  use_strand = TRUE)

}
\seealso{
Other Operating on the data regions: \code{\link{deepblue.aggregate}},
  \code{\link{deepblue.count_regions}},
  \code{\link{deepblue.extend}},
  \code{\link{deepblue.filter_regions}},
  \code{\link{deepblue.get_experiments_by_query}},
  \code{\link{deepblue.get_regions}},
  \code{\link{deepblue.input_regions}},
  \code{\link{deepblue.intersection}},
  \code{\link{deepblue.merge_queries}},
  \code{\link{deepblue.query_cache}},
  \code{\link{deepblue.query_experiment_type}},
  \code{\link{deepblue.score_matrix}},
  \code{\link{deepblue.select_annotations}},
  \code{\link{deepblue.select_experiments}},
  \code{\link{deepblue.select_regions}},
  \code{\link{deepblue.tiling_regions}}
}

