% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue.get_biosource_parents}
\alias{deepblue.get_biosource_parents}
\title{get_biosource_parents}
\usage{
deepblue.get_biosource_parents(biosource = NULL,
  user_key = deepblue.USER_KEY)
}
\arguments{
\item{biosource}{- A string (name of the biosource)}

\item{user_key}{- A string (users token key)}
}
\value{
biosources - A array (parents biosources)
}
\description{
Gets the biosources that are parents of the given biosource.
}
\seealso{
Other Set the relationship between different biosources: \code{\link{deepblue.get_biosource_children}},
  \code{\link{deepblue.get_biosource_related}},
  \code{\link{deepblue.get_biosource_synonyms}},
  \code{\link{deepblue.set_biosource_parent}},
  \code{\link{deepblue.set_biosource_synonym}}
}

