% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue.get_biosource_related}
\alias{deepblue.get_biosource_related}
\title{get_biosource_related}
\usage{
deepblue.get_biosource_related(biosource = NULL,
  user_key = deepblue.USER_KEY)
}
\arguments{
\item{biosource}{- A string (name of the biosource)}

\item{user_key}{- A string (users token key)}
}
\value{
biosources - A array (related biosources)
}
\description{
Gets biosources related to the given one. e.g. the children terms and theirs synonyms.
}
\examples{
deepblue.get_biosource_related(biosource = "Blood")

}
\seealso{
Other Set the relationship between different biosources: \code{\link{deepblue.get_biosource_children}},
  \code{\link{deepblue.get_biosource_parents}},
  \code{\link{deepblue.get_biosource_synonyms}},
  \code{\link{deepblue.set_biosource_parent}},
  \code{\link{deepblue.set_biosource_synonym}}
}

