% Generated by roxygen2: do not edit by hand
% Please edit documentation in inst/deepblue.R
\name{deepblue.get_biosource_synonyms}
\alias{deepblue.get_biosource_synonyms}
\title{get_biosource_synonyms}
\usage{
deepblue.get_biosource_synonyms(biosource = NULL,
  user_key = deepblue.USER_KEY)
}
\arguments{
\item{biosource}{- A string (name of the biosource)}

\item{user_key}{- A string (users token key)}
}
\value{
synonyms - A array (synonyms of the biosource)
}
\description{
Gets the synonyms for the biosource.
}
\seealso{
Other Set the relationship between different biosources: \code{\link{deepblue.get_biosource_children}},
  \code{\link{deepblue.get_biosource_parents}},
  \code{\link{deepblue.get_biosource_related}},
  \code{\link{deepblue.set_biosource_parent}},
  \code{\link{deepblue.set_biosource_synonym}}
}

