% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue.get_regions}
\alias{deepblue.get_regions}
\title{get_regions}
\usage{
deepblue.get_regions(query_id = NULL, output_format = NULL,
  user_key = deepblue.USER_KEY)
}
\arguments{
\item{query_id}{- A string (Query ID)}

\item{output_format}{- A string (Output format)}

\item{user_key}{- A string (users token key)}
}
\value{
request_id - A string (Request ID - Use it to retrieve the result with info() and get_request_data())
}
\description{
Send a request  to retrieve the regions for the given query in the requested BED format.
}
\examples{
\dontrun{
  deepblue.get_regions(query_id = "q123456", output_format = "CHROMOSOME,START,END",
  user_key = "my_secret_key")
}

}
\seealso{
Other Operating on the data regions: \code{\link{deepblue.aggregate}},
  \code{\link{deepblue.count_regions}},
  \code{\link{deepblue.extend}},
  \code{\link{deepblue.filter_regions}},
  \code{\link{deepblue.flank}},
  \code{\link{deepblue.get_experiments_by_query}},
  \code{\link{deepblue.input_regions}},
  \code{\link{deepblue.intersection}},
  \code{\link{deepblue.merge_queries}},
  \code{\link{deepblue.query_cache}},
  \code{\link{deepblue.query_experiment_type}},
  \code{\link{deepblue.score_matrix}},
  \code{\link{deepblue.select_annotations}},
  \code{\link{deepblue.select_experiments}},
  \code{\link{deepblue.select_regions}},
  \code{\link{deepblue.tiling_regions}}
}

