% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue.get_request_data}
\alias{deepblue.get_request_data}
\title{get_request_data}
\usage{
deepblue.get_request_data(request_id = NULL, user_key = deepblue.USER_KEY)
}
\arguments{
\item{request_id}{- A string (ID of the request)}

\item{user_key}{- A string (users token key)}
}
\value{
data - A string or a vector of string (The output can be (i) a string (get_regions, score_matrix, and count_regions), or (ii) a list of ID and names (get_experiments_by_query).)
}
\description{
Get the request data.
}
\examples{
data_id = deepblue.select_experiments(
	experiment_name="E002-H3K9ac.narrowPeak.bed", chromosome="chr1")
request_id = deepblue.get_regions(query_id =data_id, output_format = "CHROMOSOME,START,END")
deepblue.get_request_data(request_id = request_id)

}
\seealso{
Other Requests status information and results: \code{\link{deepblue.list_requests}}
}

