% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue.get_request_data_r}
\alias{deepblue.get_request_data_r}
\title{get request data r}
\usage{
deepblue.get_request_data_r(request_id, user_key = deepblue.USER_KEY,
  .defaultOpts = list(httpheader = c(`Content-Type` = "text/xml"),
  followlocation = TRUE, useragent = useragent), .curl = getCurlHandle())
}

