% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue.get_state}
\alias{deepblue.get_state}
\title{get_state}
\usage{
deepblue.get_state(data_name = NULL, user_key = deepblue.USER_KEY)
}
\arguments{
\item{data_name}{- A string (Name of the data to lookup the state for)}

\item{user_key}{- A string (users token key)}
}
\value{
data_state - A int (State of the data)
}
\description{
Returns the current state of specific data.
}
\seealso{
Other Commands for all types of data: \code{\link{deepblue.cancel_request}},
  \code{\link{deepblue.info}},
  \code{\link{deepblue.is_biosource}},
  \code{\link{deepblue.list_in_use}},
  \code{\link{deepblue.remove}},
  \code{\link{deepblue.search}}
}

