% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue.info}
\alias{deepblue.info}
\title{info}
\usage{
deepblue.info(id = NULL, user_key = deepblue.USER_KEY)
}
\arguments{
\item{id}{- A string or a vector of string (ID or an array of IDs)}

\item{user_key}{- A string (users token key)}
}
\value{
information - A array or a vector of array (List of Maps, where each map contains the info of an object.)
}
\description{
Return information for the given ID (or IDs).
}
\examples{
deepblue.info(id = "e30035")

}
\seealso{
Other Commands for all types of data: \code{\link{deepblue.cancel_request}},
  \code{\link{deepblue.is_biosource}},
  \code{\link{deepblue.list_in_use}},
  \code{\link{deepblue.remove}},
  \code{\link{deepblue.search}}
}

