% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue.input_regions}
\alias{deepblue.input_regions}
\title{input_regions}
\usage{
deepblue.input_regions(genome = NULL, region_set = NULL,
  user_key = deepblue.USER_KEY)
}
\arguments{
\item{genome}{- A string (the target genome)}

\item{region_set}{- A string (Regions in CHROMOSOME    START    END format)}

\item{user_key}{- A string (users token key)}
}
\value{
id - A string (query id)
}
\description{
Include a region set that will be used by the follow ups operations.
}
\examples{
regions_set = "chr1	28735	29810
chr1	135124	135563
chr1	327790	328229
chr1	437151	438164
chr1	449273	450544
chr1	533219	534114
chr1	544738	546649
chr1	713984	714547
chr1	762416	763445
chr1	788863	789211"
deepblue.input_regions(genome = "hg19", region_set = regions_set)

}
\seealso{
Other Operating on the data regions: \code{\link{deepblue.aggregate}},
  \code{\link{deepblue.count_regions}},
  \code{\link{deepblue.extend}},
  \code{\link{deepblue.filter_regions}},
  \code{\link{deepblue.flank}},
  \code{\link{deepblue.get_experiments_by_query}},
  \code{\link{deepblue.get_regions}},
  \code{\link{deepblue.intersection}},
  \code{\link{deepblue.merge_queries}},
  \code{\link{deepblue.query_cache}},
  \code{\link{deepblue.query_experiment_type}},
  \code{\link{deepblue.score_matrix}},
  \code{\link{deepblue.select_annotations}},
  \code{\link{deepblue.select_experiments}},
  \code{\link{deepblue.select_regions}},
  \code{\link{deepblue.tiling_regions}}
}

