% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue.is_biosource}
\alias{deepblue.is_biosource}
\title{is_biosource}
\usage{
deepblue.is_biosource(biosource_name = NULL, user_key = deepblue.USER_KEY)
}
\arguments{
\item{biosource_name}{- A string (Name of the biosource)}

\item{user_key}{- A string (users token key)}
}
\value{
information - A string or a vector of string (A string containing the biosource name)
}
\description{
Return information for the given biosource name.
}
\examples{
deepblue.is_biosource(biosource_name = "blood")

}
\seealso{
Other Commands for all types of data: \code{\link{deepblue.cancel_request}},
  \code{\link{deepblue.info}},
  \code{\link{deepblue.list_in_use}},
  \code{\link{deepblue.search}}
}

