% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue.list_annotations}
\alias{deepblue.list_annotations}
\title{list_annotations}
\usage{
deepblue.list_annotations(genome = NULL, user_key = deepblue.USER_KEY)
}
\arguments{
\item{genome}{- A string or a vector of string (the target genome)}

\item{user_key}{- A string (users token key)}
}
\value{
annotations - A array (annotation ids)
}
\description{
Lists all existing annotations.
}
\seealso{
Other Inserting and listing annotations: \code{\link{deepblue.add_annotation}},
  \code{\link{deepblue.find_pattern}}
}

