% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue.list_biosources}
\alias{deepblue.list_biosources}
\title{list_biosources}
\usage{
deepblue.list_biosources(extra_metadata = NULL,
  user_key = deepblue.USER_KEY)
}
\arguments{
\item{extra_metadata}{- A struct (Key-value that must match the biosource extra_metadata.)}

\item{user_key}{- A string (users token key)}
}
\value{
biosources - A array (biosources)
}
\description{
Lists all existing biosources.
}
\examples{
\dontrun{
  deepblue.list_biosources(extra_metadata = list(namespace = "uberon"))
}

}
\seealso{
Other Inserting and listing biosources: \code{\link{deepblue.add_biosource}},
  \code{\link{deepblue.list_similar_biosources}}
}

