% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.r
\name{deepblue.list_column_types}
\alias{deepblue.list_column_types}
\title{list_column_types}
\usage{
deepblue.list_column_types(user_key = deepblue.USER_KEY)
}
\arguments{
\item{user_key}{- A string (users token key)}
}
\value{
a vector with  ('okay',
}
\description{
Lists all available column types.

column types) or ('error', error_message)
}
\seealso{
Other Inserting and listing different column types: \code{\link{deepblue.create_column_type_calculated}},
  \code{\link{deepblue.create_column_type_category}},
  \code{\link{deepblue.create_column_type_range}},
  \code{\link{deepblue.create_column_type_simple}}
}

