% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue.list_epigenetic_marks}
\alias{deepblue.list_epigenetic_marks}
\title{list_epigenetic_marks}
\usage{
deepblue.list_epigenetic_marks(extra_metadata = NULL,
  user_key = deepblue.USER_KEY)
}
\arguments{
\item{extra_metadata}{- A struct (Key-value that must match the biosource extra_metadata.)}

\item{user_key}{- A string (users token key)}
}
\value{
epigenetic_marks - A array (epigenetic mark names)
}
\description{
Lists all existing epigenetic marks.
}
\examples{
\dontrun{
  deepblue.list_epigenetic_marks()
}

}
\seealso{
Other Inserting and listing epigenetic marks: \code{\link{deepblue.add_epigenetic_mark}},
  \code{\link{deepblue.list_similar_epigenetic_marks}}
}

