% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue.list_experiments}
\alias{deepblue.list_experiments}
\title{list_experiments}
\usage{
deepblue.list_experiments(genome = NULL, type = NULL,
  epigenetic_mark = NULL, biosource = NULL, sample = NULL,
  technique = NULL, project = NULL, user_key = deepblue.USER_KEY)
}
\arguments{
\item{genome}{- A string or a vector of string (the target genome)}

\item{type}{- A string or a vector of string (type of the experiment: peaks or signal)}

\item{epigenetic_mark}{- A string or a vector of string (name(s) of selected epigenetic mark(s))}

\item{biosource}{- A string or a vector of string (name(s) of selected biosource(s))}

\item{sample}{- A string or a vector of string (id(s) of selected sample(s))}

\item{technique}{- A string or a vector of string (name(s) of selected technique(s))}

\item{project}{- A string or a vector of string (name(s) of selected projects)}

\item{user_key}{- A string (users token key)}
}
\value{
experiments - A array (experiment names)
}
\description{
Lists all existing experiments.
}
\examples{
deepblue.list_experiments(genome = "hg19", type = "peaks",
  epigenetic_mark = "H3K27ac", biosource = "blood")

}
\seealso{
Other Inserting and listing experiments: \code{\link{deepblue.collection_experiments_count}},
  \code{\link{deepblue.faceting_experiments}},
  \code{\link{deepblue.list_recent_experiments}},
  \code{\link{deepblue.list_similar_experiments}}
}

