% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue.list_in_use}
\alias{deepblue.list_in_use}
\title{list_in_use}
\usage{
deepblue.list_in_use(controlled_vocabulary = NULL,
  user_key = deepblue.USER_KEY)
}
\arguments{
\item{controlled_vocabulary}{- A string (controlled vocabulary name)}

\item{user_key}{- A string (users token key)}
}
\value{
terms - A array (controlled_vocabulary terms with count)
}
\description{
Lists all terms from the given controlled vocabulary that are used.
}
\examples{
deepblue.list_in_use(controlled_vocabulary = "biosources")

}
\seealso{
Other Commands for all types of data: \code{\link{deepblue.cancel_request}},
  \code{\link{deepblue.info}},
  \code{\link{deepblue.is_biosource}},
  \code{\link{deepblue.search}}
}

