% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue.list_similar_epigenetic_marks}
\alias{deepblue.list_similar_epigenetic_marks}
\title{list_similar_epigenetic_marks}
\usage{
deepblue.list_similar_epigenetic_marks(name = NULL,
  user_key = deepblue.USER_KEY)
}
\arguments{
\item{name}{- A string (epigenetic mark name)}

\item{user_key}{- A string (users token key)}
}
\value{
epigenetic_marks - A array (similar epigenetic mark names)
}
\description{
Lists all epigenetic marks similar to the one provided.
}
\seealso{
Other Inserting and listing epigenetic marks: \code{\link{deepblue.add_epigenetic_mark}},
  \code{\link{deepblue.list_epigenetic_marks}}
}

