% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.r
\name{deepblue.list_similar_experiments}
\alias{deepblue.list_similar_experiments}
\title{list_similar_experiments}
\usage{
deepblue.list_similar_experiments(name = NULL, genome = NULL,
  user_key = deepblue.USER_KEY)
}
\arguments{
\item{name}{- A string (experiment name)}

\item{genome}{- A string or a vector of string (the target genome)}

\item{user_key}{- A string (users token key)}
}
\value{
a vector with  ('okay',
}
\description{
Lists all experiments similar to the one provided.

similar experiment names) or ('error', error_message)
}
\seealso{
Other Inserting and listing experiments: \code{\link{deepblue.add_experiment}},
  \code{\link{deepblue.faceting_experiments}},
  \code{\link{deepblue.list_experiments}},
  \code{\link{deepblue.list_recent_experiments}}
}

