% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue.list_similar_genomes}
\alias{deepblue.list_similar_genomes}
\title{list_similar_genomes}
\usage{
deepblue.list_similar_genomes(name = NULL, user_key = deepblue.USER_KEY)
}
\arguments{
\item{name}{- A string (genome name)}

\item{user_key}{- A string (users token key)}
}
\value{
genomes - A array (similar genome names)
}
\description{
Lists all genomes similar to the one provided.
}
\examples{
\dontrun{
  deepblue.list_similar_genomes(name = "grc")
}

}
\seealso{
Other Inserting and listing genomes: \code{\link{deepblue.add_genome}},
  \code{\link{deepblue.chromosomes}},
  \code{\link{deepblue.list_genomes}},
  \code{\link{deepblue.upload_chromosome}}
}

