% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue.list_similar_projects}
\alias{deepblue.list_similar_projects}
\title{list_similar_projects}
\usage{
deepblue.list_similar_projects(name = NULL, user_key = deepblue.USER_KEY)
}
\arguments{
\item{name}{- A string (project name)}

\item{user_key}{- A string (users token key)}
}
\value{
projects - A array (similar project names)
}
\description{
Lists all projects similar to the one provided.
}
\examples{
deepblue.list_similar_projects(name = "BLUEPRINT")

}
\seealso{
Other Inserting and listing projects: \code{\link{deepblue.list_projects}}
}

