% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.r
\name{deepblue.list_techniques}
\alias{deepblue.list_techniques}
\title{list_techniques}
\usage{
deepblue.list_techniques(user_key = deepblue.USER_KEY)
}
\arguments{
\item{user_key}{- A string (users token key)}
}
\value{
a vector with  ('okay',
}
\description{
Lists all existing techniques.

techniques) or ('error', error_message)
}
\seealso{
Other Inserting and listing techniques: \code{\link{deepblue.add_technique}},
  \code{\link{deepblue.list_similar_techniques}}
}

