% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue.merge_queries}
\alias{deepblue.merge_queries}
\title{merge_queries}
\usage{
deepblue.merge_queries(query_a_id = NULL, query_b_id = NULL,
  user_key = deepblue.USER_KEY)
}
\arguments{
\item{query_a_id}{- A string (id of the first query)}

\item{query_b_id}{- A string (id of the second query)}

\item{user_key}{- A string (users token key)}
}
\value{
id - A string (new query id)
}
\description{
Merges the regions of the given queries.
}
\examples{
\dontrun{
  deepblue.merge_queries(query_a_id = query_a, query_b_id = query_a,
  user_key = "my_secret_key")
}

}
\seealso{
Other Operating on the data regions: \code{\link{deepblue.aggregate}},
  \code{\link{deepblue.count_regions}},
  \code{\link{deepblue.extend}},
  \code{\link{deepblue.filter_regions}},
  \code{\link{deepblue.flank}},
  \code{\link{deepblue.get_experiments_by_query}},
  \code{\link{deepblue.get_regions}},
  \code{\link{deepblue.input_regions}},
  \code{\link{deepblue.intersection}},
  \code{\link{deepblue.query_cache}},
  \code{\link{deepblue.query_experiment_type}},
  \code{\link{deepblue.score_matrix}},
  \code{\link{deepblue.select_annotations}},
  \code{\link{deepblue.select_experiments}},
  \code{\link{deepblue.select_regions}},
  \code{\link{deepblue.tiling_regions}}
}

