% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalFunctions.R
\name{deepblue.parse_gtf}
\alias{deepblue.parse_gtf}
\title{deepblue.parse_gtf}
\usage{
deepblue.parse_gtf(all_gtf)
}
\arguments{
\item{all_gtf}{a character vector of GTF attributes for a single region.}
}
\description{
Parse the GTF semicolon separated attributes into a data frame
}
\keyword{internal}

