% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue.query_cache}
\alias{deepblue.query_cache}
\title{query_cache}
\usage{
deepblue.query_cache(query_id = NULL, cache = NULL,
  user_key = deepblue.USER_KEY)
}
\arguments{
\item{query_id}{- A string (query ID)}

\item{cache}{- A boolean (set or unset this query caching)}

\item{user_key}{- A string (users token key)}
}
\value{
information - A string (New query ID.)
}
\description{
Return information for the given ID (or IDs).
}
\seealso{
Other Operating on the data regions: \code{\link{deepblue.aggregate}},
  \code{\link{deepblue.count_regions}},
  \code{\link{deepblue.extend}},
  \code{\link{deepblue.filter_regions}},
  \code{\link{deepblue.flank}},
  \code{\link{deepblue.get_experiments_by_query}},
  \code{\link{deepblue.get_regions}},
  \code{\link{deepblue.input_regions}},
  \code{\link{deepblue.intersection}},
  \code{\link{deepblue.merge_queries}},
  \code{\link{deepblue.query_experiment_type}},
  \code{\link{deepblue.score_matrix}},
  \code{\link{deepblue.select_annotations}},
  \code{\link{deepblue.select_experiments}},
  \code{\link{deepblue.select_regions}},
  \code{\link{deepblue.tiling_regions}}
}

