% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue.query_experiment_type}
\alias{deepblue.query_experiment_type}
\title{query_experiment_type}
\usage{
deepblue.query_experiment_type(query_id = NULL, type = NULL,
  user_key = deepblue.USER_KEY)
}
\arguments{
\item{query_id}{- A string (query ID)}

\item{type}{- A string (experiment type (peaks or signal))}

\item{user_key}{- A string (users token key)}
}
\value{
information - A string (New query ID.)
}
\description{
Return information for the given ID (or IDs).
}
\examples{
deepblue.query_experiment_type(query_id = "q1234", type = "signal",
  user_key = "my_secret_key")

}
\seealso{
Other Operating on the data regions: \code{\link{deepblue.aggregate}},
  \code{\link{deepblue.count_regions}},
  \code{\link{deepblue.extend}},
  \code{\link{deepblue.filter_regions}},
  \code{\link{deepblue.flank}},
  \code{\link{deepblue.get_experiments_by_query}},
  \code{\link{deepblue.get_regions}},
  \code{\link{deepblue.input_regions}},
  \code{\link{deepblue.intersection}},
  \code{\link{deepblue.merge_queries}},
  \code{\link{deepblue.query_cache}},
  \code{\link{deepblue.score_matrix}},
  \code{\link{deepblue.select_annotations}},
  \code{\link{deepblue.select_experiments}},
  \code{\link{deepblue.select_regions}},
  \code{\link{deepblue.tiling_regions}}
}

