% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue.select_annotations}
\alias{deepblue.select_annotations}
\title{select_annotations}
\usage{
deepblue.select_annotations(annotation_name = NULL, genome = NULL,
  chromosome = NULL, start = NULL, end = NULL,
  user_key = deepblue.USER_KEY)
}
\arguments{
\item{annotation_name}{- A string or a vector of string (name(s) of selected annotation(s))}

\item{genome}{- A string or a vector of string (the target genome)}

\item{chromosome}{- A string or a vector of string (chromosome name(s))}

\item{start}{- A int (minimum start region)}

\item{end}{- A int (maximum end region)}

\item{user_key}{- A string (users token key)}
}
\value{
id - A string (query id)
}
\description{
Selects annotation regions matching the given parameters.
}
\examples{
deepblue.select_annotations(annotation_name = "Cpg Islands", genome = "hg19",
  chromosome = "chr1", start = 0, end = 2000000)

}
\seealso{
Other Operating on the data regions: \code{\link{deepblue.aggregate}},
  \code{\link{deepblue.count_regions}},
  \code{\link{deepblue.extend}},
  \code{\link{deepblue.filter_regions}},
  \code{\link{deepblue.flank}},
  \code{\link{deepblue.get_experiments_by_query}},
  \code{\link{deepblue.get_regions}},
  \code{\link{deepblue.input_regions}},
  \code{\link{deepblue.intersection}},
  \code{\link{deepblue.merge_queries}},
  \code{\link{deepblue.query_cache}},
  \code{\link{deepblue.query_experiment_type}},
  \code{\link{deepblue.score_matrix}},
  \code{\link{deepblue.select_experiments}},
  \code{\link{deepblue.select_regions}},
  \code{\link{deepblue.tiling_regions}}
}

