% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.r
\name{deepblue.select_genes}
\alias{deepblue.select_genes}
\title{select_genes}
\usage{
deepblue.select_genes(genes_name = NULL, gene_set = NULL,
  user_key = deepblue.USER_KEY)
}
\arguments{
\item{genes_name}{- A string or a vector of string (genes(s) (ENSB ID or ENSB name))}

\item{gene_set}{- A string (gene set name)}

\item{user_key}{- A string (users token key)}
}
\value{
a vector with  ('okay',
}
\description{
Selects genes as regions.

query id) or ('error', error_message)
}
\seealso{
Other Operations on gene sets and genes identifiers: \code{\link{deepblue.add_gene_set}}
}

