% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue.select_regions}
\alias{deepblue.select_regions}
\title{select_regions}
\usage{
deepblue.select_regions(experiment_name = NULL, genome = NULL,
  epigenetic_mark = NULL, sample_id = NULL, technique = NULL,
  project = NULL, chromosome = NULL, start = NULL, end = NULL,
  user_key = deepblue.USER_KEY)
}
\arguments{
\item{experiment_name}{- A string or a vector of string (name(s) of selected experiment(s))}

\item{genome}{- A string or a vector of string (the target genome)}

\item{epigenetic_mark}{- A string or a vector of string (name(s) of selected epigenetic mark(s))}

\item{sample_id}{- A string or a vector of string (id(s) of selected sample(s))}

\item{technique}{- A string or a vector of string (name(s) of selected technique(es))}

\item{project}{- A string or a vector of string (name(s) of selected projects)}

\item{chromosome}{- A string or a vector of string (chromosome name(s))}

\item{start}{- A int (minimum start region)}

\item{end}{- A int (maximum end region)}

\item{user_key}{- A string (users token key)}
}
\value{
id - A string (query id)
}
\description{
Selects experiment regions matching the given parameters.
}
\examples{
deepblue.select_regions(genome="hg19", epigenetic_mark = "H3K27ac",
  project = " BLUEPRINT Epigenome")

}
\seealso{
Other Operating on the data regions: \code{\link{deepblue.aggregate}},
  \code{\link{deepblue.count_regions}},
  \code{\link{deepblue.extend}},
  \code{\link{deepblue.filter_regions}},
  \code{\link{deepblue.flank}},
  \code{\link{deepblue.get_experiments_by_query}},
  \code{\link{deepblue.get_regions}},
  \code{\link{deepblue.input_regions}},
  \code{\link{deepblue.intersection}},
  \code{\link{deepblue.merge_queries}},
  \code{\link{deepblue.query_cache}},
  \code{\link{deepblue.query_experiment_type}},
  \code{\link{deepblue.score_matrix}},
  \code{\link{deepblue.select_annotations}},
  \code{\link{deepblue.select_experiments}},
  \code{\link{deepblue.tiling_regions}}
}

