% Generated by roxygen2: do not edit by hand
% Please edit documentation in inst/deepblue.R
\name{deepblue.set_biosource_parent}
\alias{deepblue.set_biosource_parent}
\title{set_biosource_parent}
\usage{
deepblue.set_biosource_parent(parent = NULL, child = NULL,
  user_key = deepblue.USER_KEY)
}
\arguments{
\item{parent}{- A string (parent)}

\item{child}{- A string (child)}

\item{user_key}{- A string (users token key)}
}
\description{
Sets a biosource parent.
}
\seealso{
Other Set the relationship between different biosources: \code{\link{deepblue.get_biosource_children}},
  \code{\link{deepblue.get_biosource_parents}},
  \code{\link{deepblue.get_biosource_related}},
  \code{\link{deepblue.get_biosource_synonyms}},
  \code{\link{deepblue.set_biosource_synonym}}
}

