% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue.set_project_public}
\alias{deepblue.set_project_public}
\title{set_project_public}
\usage{
deepblue.set_project_public(project = NULL, set = NULL,
  user_key = deepblue.USER_KEY)
}
\arguments{
\item{project}{- A string (Project name or ID)}

\item{set}{- A boolean (True to set the project as public of false for unset)}

\item{user_key}{- A string (users token key)}
}
\value{
id - A string (id of the project)
}
\description{
Set a project as public. You must be the project owner to perform this operation.
}
\examples{
\dontrun{
  deepblue.set_project_public(project = "My secret project", set = TRUE,
  user_key = "my_secret_key")
}

}
\seealso{
Other Inserting and listing projects: \code{\link{deepblue.add_project}},
  \code{\link{deepblue.add_user_to_project}},
  \code{\link{deepblue.list_projects}},
  \code{\link{deepblue.list_similar_projects}}
}

