% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalFunctions.R
\name{deepblue.switch_get_request_data}
\alias{deepblue.switch_get_request_data}
\alias{get_request_data}
\title{switch_get_request_data}
\usage{
deepblue.switch_get_request_data(request_id, user_key = deepblue.USER_KEY)
}
\arguments{
\item{request_id}{The request command generated
with deepblue.get_request_data}

\item{user_key}{the user_key used to submit the request.}
}
\description{
Download the requested data from DeepBlue.
Can be either region sets,
a region count, a list of experiments, or a score matrix.
}
\keyword{internal}

