% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue.tiling_regions}
\alias{deepblue.tiling_regions}
\title{tiling_regions}
\usage{
deepblue.tiling_regions(size = NULL, genome = NULL, chromosome = NULL,
  user_key = deepblue.USER_KEY)
}
\arguments{
\item{size}{- A int (tiling size)}

\item{genome}{- A string (the target genome)}

\item{chromosome}{- A string or a vector of string (chromosome name(s))}

\item{user_key}{- A string (users token key)}
}
\value{
id - A string (query id)
}
\description{
Creates regions with the tiling size over the chromosomes.
}
\examples{
deepblue.tiling_regions(
    size = 10000,
    genome = "hg19",
    chromosome = "chr1")

}
\seealso{
Other Operating on the data regions: \code{\link{deepblue.aggregate}},
  \code{\link{deepblue.count_regions}},
  \code{\link{deepblue.extend}},
  \code{\link{deepblue.filter_regions}},
  \code{\link{deepblue.flank}},
  \code{\link{deepblue.get_experiments_by_query}},
  \code{\link{deepblue.get_regions}},
  \code{\link{deepblue.input_regions}},
  \code{\link{deepblue.intersection}},
  \code{\link{deepblue.merge_queries}},
  \code{\link{deepblue.query_cache}},
  \code{\link{deepblue.query_experiment_type}},
  \code{\link{deepblue.score_matrix}},
  \code{\link{deepblue.select_annotations}},
  \code{\link{deepblue.select_experiments}},
  \code{\link{deepblue.select_regions}}
}

