% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.r
\name{deepblue.upload_chromosome}
\alias{deepblue.upload_chromosome}
\title{upload_chromosome}
\usage{
deepblue.upload_chromosome(genome = NULL, chromosome = NULL, data = NULL,
  user_key = deepblue.USER_KEY)
}
\arguments{
\item{genome}{- A string (the target genome)}

\item{chromosome}{- A string (chromosome name)}

\item{data}{- A string (chromosome sequence data)}

\item{user_key}{- A string (users token key)}
}
\value{
a vector with  ('okay', ) or ('error', error_message)
}
\description{
Uploads the sequence data of the chromosome.
}
\seealso{
Other Inserting and listing genomes: \code{\link{deepblue.add_genome}},
  \code{\link{deepblue.chromosomes}},
  \code{\link{deepblue.list_genomes}},
  \code{\link{deepblue.list_similar_genomes}}
}

