% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\docType{data}
\name{deepblue.URL}
\alias{deepblue.URL}
\title{deepblue URL}
\format{An object of class \code{character} of length 1.}
\usage{
deepblue.URL
}
\value{
URL to access DeepBlue Data Server
}
\description{
Location of the DeepBlue XML-RPC server
}
\keyword{internal}

