% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalFunctions.R
\name{deepblue.wait_request}
\alias{deepblue.wait_request}
\title{deepblue.wait_request}
\usage{
deepblue.wait_request(request_id, sleep_time = 1,
  user_key = deepblue.USER_KEY)
}
\arguments{
\item{request_id}{A string with the request_id}

\item{sleep_time}{An integer with default value 1s}

\item{user_key}{A string}
}
\description{
Process the user request. Takes in three parameters; requested regions, sleep time, and user key.
}

