% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{deepblue_batch_export_results}
\alias{deepblue_batch_export_results}
\title{batch_export_results}
\usage{
deepblue_batch_export_results(requests, target.directory = NULL,
  suffix = "result", prefix = "DeepBlue", sleep.time = 1,
  bed.format = TRUE, user_key = deepblue_options("user_key"))
}
\arguments{
\item{requests}{A list of request objects}

\item{target.directory}{Where the results should be saved}

\item{suffix}{File names suffix}

\item{prefix}{File names prefix}

\item{sleep.time}{How long this function will wait after the requests
verification}

\item{bed.format}{whether to store the results as BED files or tab delimited.}

\item{user_key}{A string used to authenticate the user}
}
\value{
A list containing the requests IDs data
}
\description{
Write results from DeepBlue to disk as they become available
}
\examples{
data_id = deepblue_select_experiments(
experiment_name="E002-H3K9ac.narrowPeak.bed", chromosome="chr1")
request_id = deepblue_get_regions(query_id =data_id,
  output_format = "CHROMOSOME,START,END")
request_data = deepblue_batch_export_results(list(request_id))
}
