% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching.R
\name{deepblue_cache_status}
\alias{deepblue_cache_status}
\title{Report on the cache size and status}
\usage{
deepblue_cache_status()
}
\value{
cache size in byte
}
\description{
Report on the cache size and status
}
\examples{
deepblue_cache_status()
}
