% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue_collection_experiments_count}
\alias{deepblue_collection_experiments_count}
\title{collection_experiments_count}
\usage{
deepblue_collection_experiments_count(controlled_vocabulary = NULL,
  genome = NULL, type = NULL, epigenetic_mark = NULL, biosource = NULL,
  sample = NULL, technique = NULL, project = NULL,
  user_key = deepblue_options("user_key"))
}
\arguments{
\item{controlled_vocabulary}{- A string (controlled vocabulary name)}

\item{genome}{- A string or a vector of string (the target genome)}

\item{type}{- A string or a vector of string (type of the experiment: peaks or signal)}

\item{epigenetic_mark}{- A string or a vector of string (name(s) of selected epigenetic mark(s))}

\item{biosource}{- A string or a vector of string (name(s) of selected biosource(s))}

\item{sample}{- A string or a vector of string (id(s) of selected sample(s))}

\item{technique}{- A string or a vector of string (name(s) of selected technique(s))}

\item{project}{- A string or a vector of string (name(s) of selected projects)}

\item{user_key}{- A string (users token key)}
}
\value{
terms - A array (controlled_vocabulary terms with count)
}
\description{
Count the number of experiments that match the selection criteria in each term of the selected controlled_vocabulary. The selection can be achieved through specifying a list of BioSources, experimental Techniques, Epigenetic Marks, Samples or Projects.
}
\examples{
deepblue_collection_experiments_count(
	controlled_vocabulary="epigenetic_marks",
    genome = "hg19", type = "peaks",
    biosource = "blood")

}
\seealso{
Other Inserting and listing experiments: \code{\link{deepblue_faceting_experiments}},
  \code{\link{deepblue_list_experiments}},
  \code{\link{deepblue_list_recent_experiments}},
  \code{\link{deepblue_list_similar_experiments}},
  \code{\link{deepblue_preview_experiment}}
}
