% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalFunctions.R
\name{deepblue_column_types}
\alias{deepblue_column_types}
\title{get columns}
\usage{
deepblue_column_types()
}
\value{
Dictionary will all column names and types
}
\description{
Load the column types from DeepBlue
}
\keyword{internal}
