% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue_commands}
\alias{deepblue_commands}
\title{commands}
\usage{
deepblue_commands()
}
\value{
commands - A struct (command descriptions)
}
\description{
List all available DeepBlue commands.
}
\examples{
deepblue_commands()

}
\seealso{
Other Checking DeepBlue status: \code{\link{deepblue_echo}}
}
