% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalFunctions.R
\name{deepblue_convert_to_df}
\alias{deepblue_convert_to_df}
\title{convert_to_df}
\usage{
deepblue_convert_to_df(file_to_parse, request_info, dict = col_dict)
}
\arguments{
\item{request_info}{The request information returned by DeepBlue}

\item{dict}{The data structure that contains the DeepBlue columns types}

\item{string_to_parse}{A string}
}
\value{
regions A data frame
}
\description{
save output in a data frame for further processing.
Expects two parameters; the output string from method
deepblue_get_request_data and request information from process_request
}
\keyword{internal}
