% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalFunctions.R
\name{deepblue_convert_to_grange}
\alias{deepblue_convert_to_grange}
\title{convert_to_grange}
\usage{
deepblue_convert_to_grange(df = NULL)
}
\arguments{
\item{df}{A data frame}
}
\value{
region_gr A GRanges object
}
\description{
Converts the requested data into GRanges object.
Expects one input; A dataframe with requested data.
}
\seealso{
\code{\link[GenomicRanges]{makeGRangesFromDataFrame}}
}
\keyword{internal}
