% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue_count_gene_ontology_terms}
\alias{deepblue_count_gene_ontology_terms}
\title{count_gene_ontology_terms}
\usage{
deepblue_count_gene_ontology_terms(genes = NULL, go_terms = NULL,
  chromosome = NULL, start = NULL, end = NULL, gene_model = NULL,
  user_key = deepblue_options("user_key"))
}
\arguments{
\item{genes}{- A string or a vector of string (Name(s) or ENSEMBL ID (ENSGXXXXXXXXXXX.X ) of the gene(s).)}

\item{go_terms}{- A string or a vector of string (gene ontology terms - ID or label)}

\item{chromosome}{- A string or a vector of string (chromosome name(s))}

\item{start}{- A int (minimum start region)}

\item{end}{- A int (maximum end region)}

\item{gene_model}{- A string (the gene model)}

\item{user_key}{- A string (users token key)}
}
\value{
faceting - A struct (Map with the mandatory fields of the experiments metadata, where each contains a list of terms that appears.)
}
\description{
Summarize the controlled_vocabulary fields, from experiments that match the selection criteria. It is similar to the 'collection_experiments_count' command, but this command return the summarization for all controlled_vocabulary terms.
}
\examples{
gene_names = c('CCR1', 'CD164', 'CD1D', 'CD2', 'CD34', 'CD3G', 'CD44')
deepblue_count_gene_ontology_terms (genes = gene_names, gene_model = "gencode v23")


}
\seealso{
Other Gene models and genes identifiers: \code{\link{deepblue_list_gene_models}},
  \code{\link{deepblue_list_genes}},
  \code{\link{deepblue_select_genes}}
}
