% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\docType{data}
\name{deepblue_debug_VERBOSE}
\alias{deepblue_debug_VERBOSE}
\title{Verbose}
\format{An object of class \code{logical} of length 1.}
\usage{
deepblue_debug_VERBOSE
}
\value{
if this package must print some debugging outputs
}
\description{
Show or hide debugging messages
}
\keyword{internal}

