% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching.R
\name{deepblue_delete_request_from_cache}
\alias{deepblue_delete_request_from_cache}
\title{Delete a specific request from the cache}
\usage{
deepblue_delete_request_from_cache(request_id)
}
\arguments{
\item{request_id}{the request to delete from the cache}
}
\value{
TRUE if the request was successfully deleted, FALSE otherwise
}
\description{
Delete a specific request from the cache
}
\examples{
deepblue_delete_request_from_cache("non-existing-request-id")
# returns FALSE
}
