% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{deepblue_diff}
\alias{deepblue_diff}
\title{diff}
\usage{
deepblue_diff(id1, id2, user_key = deepblue_options("user_key"))
}
\arguments{
\item{id1}{- A DeepBlue id}

\item{id2}{- Another DeepBlue id}

\item{user_key}{- A string (users token key)}
}
\value{
None
}
\description{
A utility command that creates a diff view of info
for two DeepBlue ids
}
\examples{
deepblue_diff(
    id1 = "e16918",
    id2 = "e16919")

}
\seealso{
Other Utilities for information processing: \code{\link{deepblue_select_column}}
}
