% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue_distinct_column_values}
\alias{deepblue_distinct_column_values}
\title{distinct_column_values}
\usage{
deepblue_distinct_column_values(query_id = NULL, field = NULL,
  user_key = deepblue_options("user_key"))
}
\arguments{
\item{query_id}{- A string (Query ID)}

\item{field}{- A string (field that is filtered by)}

\item{user_key}{- A string (users token key)}
}
\value{
id - A string (id of filtered query)
}
\description{
Obtain the distict values of the field.
}
\examples{
css_experiment <- deepblue_select_experiments ( "wgEncodeBroadHmmK562HMM")
distinct_names_request <- deepblue_distinct_column_values (css_experiment, "NAME")

}
\seealso{
Other Operating on the data regions: \code{\link{deepblue_aggregate}},
  \code{\link{deepblue_binning}},
  \code{\link{deepblue_count_regions}},
  \code{\link{deepblue_coverage}},
  \code{\link{deepblue_extend}},
  \code{\link{deepblue_filter_regions}},
  \code{\link{deepblue_flank}},
  \code{\link{deepblue_get_experiments_by_query}},
  \code{\link{deepblue_get_regions}},
  \code{\link{deepblue_input_regions}},
  \code{\link{deepblue_intersection}},
  \code{\link{deepblue_merge_queries}},
  \code{\link{deepblue_overlap}},
  \code{\link{deepblue_query_cache}},
  \code{\link{deepblue_query_experiment_type}},
  \code{\link{deepblue_score_matrix}},
  \code{\link{deepblue_select_annotations}},
  \code{\link{deepblue_select_experiments}},
  \code{\link{deepblue_select_regions}},
  \code{\link{deepblue_tiling_regions}}
}
