% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_DeepBlueCommand.R
\docType{methods}
\name{deepblue_download_request_data,DeepBlueCommand-method}
\alias{deepblue_download_request_data,DeepBlueCommand-method}
\title{deepblue_download_request_data}
\usage{
\S4method{deepblue_download_request_data}{DeepBlueCommand}(request_id)
}
\arguments{
\item{request_id}{DeepBlueCommand object}
}
\value{
grange_regions Final output in GRanges format
}
\description{
Returns the requested data as the expected type object.
Expects two input parameters; Request information and
user key. It depends on outputs from several functions, namely;
deepblue_get_request_data, convert_to_df, and convert_to_grange.
}
