% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalFunctions.R
\name{deepblue_download_request_data}
\alias{deepblue_download_request_data}
\title{deepblue_download_request_data}
\usage{
deepblue_download_request_data(request_id,
  user_key = deepblue_options("user_key"),
  force_download = deepblue_options("force_download"),
  do_not_cache = deepblue_options("do_not_cache"))
}
\arguments{
\item{request_id}{- Id of the request that will be downloaded}

\item{user_key}{A string}

\item{force_download}{forces DeepBlueR to download the request overwriting
any results that might already be in the cache}

\item{do_not_cache}{whether to use local caching of requests}
}
\value{
grange_regions Final output in GRanges format or as data frame
}
\description{
Returns the requested data as the expected type object.
Expects two input parameters; Request information and
user key. It depends on outputs from several functions, namely;
deepblue_get_request_data, convert_to_df, and convert_to_grange.
}
\examples{
data_id = deepblue_select_experiments(
experiment_name="E002-H3K9ac.narrowPeak.bed", chromosome="chr1")
request_id = deepblue_get_regions(query_id =data_id,
  output_format = "CHROMOSOME,START,END")
request_data = deepblue_download_request_data(request_id)
}

