% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue_enrich_regions_go_terms}
\alias{deepblue_enrich_regions_go_terms}
\title{enrich_regions_go_terms}
\usage{
deepblue_enrich_regions_go_terms(query_id = NULL, gene_model = NULL,
  user_key = deepblue_options("user_key"))
}
\arguments{
\item{query_id}{- A string (Query ID)}

\item{gene_model}{- A string (the gene model)}

\item{user_key}{- A string (users token key)}
}
\value{
request_id - A string (Request ID - Use it to retrieve the result with info() and get_request_data(). The result is a list containing the GO terms that overlap with the query_id regions.)
}
\description{
Enrich the regions based on Gene Ontology terms.
}
\examples{
data_id = deepblue_select_experiments(
    experiment_name="E002-H3K9ac.narrowPeak.bed")

filtered_id = deepblue_filter_regions(query_id = data_id,
    field = "VALUE",
    operation = ">",
    value = "100",
    type = "number",
    user_key = "anonymous_key")

deepblue_enrich_regions_go_terms(query_id = filtered_id,
   gene_model = "gencode v23")

}
\seealso{
Other Enrich the genome regions: \code{\link{deepblue_enrich_region_overlap}}
}
