% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{deepblue_export_meta_data}
\alias{deepblue_export_meta_data}
\title{Export meta data as tab delimited file}
\usage{
deepblue_export_meta_data(ids, target.directory = "./", file.name,
  user_key = deepblue_options("user_key"))
}
\arguments{
\item{ids}{an id or a list of DeepBlue ids}

\item{target.directory}{where the meta data should be stored}

\item{file.name}{name of the file}

\item{user_key}{DeepBlue user key}
}
\value{
return value of write.table
}
\description{
Export meta data as tab delimited file
}
\examples{
deepblue_export_meta_data(list("e30035", "e30036"),
file.name = "test_export",
target.directory = tempdir())
}
