% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{deepblue_export_tab}
\alias{deepblue_export_tab}
\title{Export a DeepBlue result as ordinary tab delimited file}
\usage{
deepblue_export_tab(result, target.directory = "./", file.name)
}
\arguments{
\item{result}{A result from a DeepBlue request such as a set of genomic
regions.}

\item{target.directory}{The directory to save the file to}

\item{file.name}{The name of the file without suffix}
}
\value{
return value of write.table
}
\description{
Export a DeepBlue result as ordinary tab delimited file
}
\examples{
query_id = deepblue_select_experiments (
experiment=c("GC_T14_10.CPG_methylation_calls.bs_call.GRCh38.20160531.wig"),
chromosome="chr1", start=0, end=50000000)
cpg_islands = deepblue_select_annotations(annotation_name="CpG Islands",
genome="GRCh38", chromosome="chr1", start=0, end=50000000)
overlapped = deepblue_aggregate (data_id=query_id, ranges_id=cpg_islands,
                                column="VALUE" )
request_id = deepblue_get_regions(query_id=overlapped,
                                output_format=
                                     "CHROMOSOME,START,END,@AGG.MIN,@AGG.MAX,@AGG.MEAN,@AGG.VAR")
regions = deepblue_download_request_data(request_id=request_id)
temp_dir = tempdir()
deepblue_export_tab(regions, target.directory = temp_dir,
                  file.name = "GC_T14_10.CpG_islands")
}
