% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue_extend}
\alias{deepblue_extend}
\title{extend}
\usage{
deepblue_extend(query_id = NULL, length = NULL, direction = NULL,
  use_strand = NULL, user_key = deepblue_options("user_key"))
}
\arguments{
\item{query_id}{- A string (id of the query that contains the regions)}

\item{length}{- A int (The new region length)}

\item{direction}{- A string (The direction that the region will be extended: 'BACKWARD', 'FORWARD', 'BOTH'. (Empty value will be used for both direction.)}

\item{use_strand}{- A boolean (Use the region column STRAND to define the region direction)}

\item{user_key}{- A string (users token key)}
}
\value{
id - A string (id of the new query)
}
\description{
Extend the genomic regions included in the query. It is possible to extend downstream, upstream or in both directions.
}
\examples{
annotation_id = deepblue_select_annotations(
    annotation_name="CpG Islands",
    genome="hg19", chromosome="chr1")
deepblue_extend(query_id = annotation_id,
    length = 2000, direction = "BOTH",
    use_strand = TRUE)

}
\seealso{
Other Operating on the data regions: \code{\link{deepblue_aggregate}},
  \code{\link{deepblue_binning}},
  \code{\link{deepblue_count_regions}},
  \code{\link{deepblue_coverage}},
  \code{\link{deepblue_filter_regions}},
  \code{\link{deepblue_flank}},
  \code{\link{deepblue_get_experiments_by_query}},
  \code{\link{deepblue_get_regions}},
  \code{\link{deepblue_input_regions}},
  \code{\link{deepblue_intersection}},
  \code{\link{deepblue_merge_queries}},
  \code{\link{deepblue_overlap}},
  \code{\link{deepblue_query_cache}},
  \code{\link{deepblue_query_experiment_type}},
  \code{\link{deepblue_score_matrix}},
  \code{\link{deepblue_select_annotations}},
  \code{\link{deepblue_select_experiments}},
  \code{\link{deepblue_select_regions}},
  \code{\link{deepblue_tiling_regions}}
}

