% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue_extract_ids}
\alias{deepblue_extract_ids}
\title{extract_ids}
\usage{
deepblue_extract_ids(list = NULL)
}
\arguments{
\item{list}{- A array (list of lists of IDs and Names)}
}
\value{
ids - A array (list containing the extracted IDs)
}
\description{
Extract the names from a list of ID and Names.
}
\examples{
deepblue_extract_ids(
    list = list(list("a124", "Annotation 1"),
    list("a1235", "Annotation 2")))

}
\seealso{
Other Utilities for connecting operations: \code{\link{deepblue_extract_names}}
}

