% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue_faceting_experiments}
\alias{deepblue_faceting_experiments}
\title{faceting_experiments}
\usage{
deepblue_faceting_experiments(genome = NULL, type = NULL,
  epigenetic_mark = NULL, biosource = NULL, sample = NULL,
  technique = NULL, project = NULL,
  user_key = deepblue_options("user_key"))
}
\arguments{
\item{genome}{- A string or a vector of string (the target genome)}

\item{type}{- A string or a vector of string (type of the experiment: peaks or signal)}

\item{epigenetic_mark}{- A string or a vector of string (name(s) of selected epigenetic mark(s))}

\item{biosource}{- A string or a vector of string (name(s) of selected biosource(s))}

\item{sample}{- A string or a vector of string (id(s) of selected sample(s))}

\item{technique}{- A string or a vector of string (name(s) of selected technique(s))}

\item{project}{- A string or a vector of string (name(s) of selected projects)}

\item{user_key}{- A string (users token key)}
}
\value{
faceting - A struct (Map with the mandatory fields of the experiments metadata, where each contains a list of terms that appears.)
}
\description{
Summarize the controlled_vocabulary fields, from experiments that match the selection criteria. It is similar to the 'collection_experiments_count' command, but this command return the summarization for all controlled_vocabulary terms.
}
\examples{
deepblue_faceting_experiments(genome = "hg19",
    type = "peaks",
    biosource = "blood")

}
\seealso{
Other Inserting and listing experiments: \code{\link{deepblue_collection_experiments_count}},
  \code{\link{deepblue_list_experiments}},
  \code{\link{deepblue_list_recent_experiments}},
  \code{\link{deepblue_list_similar_experiments}},
  \code{\link{deepblue_preview_experiment}}
}
