% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue_find_motif}
\alias{deepblue_find_motif}
\title{find_motif}
\usage{
deepblue_find_motif(motif = NULL, genome = NULL, chromosomes = NULL,
  start = NULL, end = NULL, overlap = NULL,
  user_key = deepblue_options("user_key"))
}
\arguments{
\item{motif}{- A string (motif (PERL regular expression))}

\item{genome}{- A string (the target genome)}

\item{chromosomes}{- A string or a vector of string (chromosome name(s))}

\item{start}{- A int (minimum start region)}

\item{end}{- A int (maximum end region)}

\item{overlap}{- A boolean (if the matching should do overlap search)}

\item{user_key}{- A string (users token key)}
}
\value{
id - A string (id of the annotation that contains the positions of the given motif)
}
\description{
Find genomic regions based on a given motif that appears in the genomic sequence.
}
\examples{
deepblue_find_motif(motif = "C[GT]+C", chromosomes=c("chr11", "chr12"),
    genome = "hg19", overlap = FALSE)

}
\seealso{
Other Inserting and listing annotations: \code{\link{deepblue_list_annotations}}
}
