% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue_find_pattern}
\alias{deepblue_find_pattern}
\title{find_pattern}
\usage{
deepblue_find_pattern(pattern = NULL, genome = NULL, overlap = NULL,
  user_key = deepblue_options("user_key"))
}
\arguments{
\item{pattern}{- A string (pattern (PERL regular expression))}

\item{genome}{- A string (the target genome)}

\item{overlap}{- A boolean (if the matching should do overlap search)}

\item{user_key}{- A string (users token key)}
}
\value{
id - A string (id of the annotation that contains the positions of the given pattern)
}
\description{
Generate a custom annotation of genomic regions based on a given pattern that appears in the genomic sequence. Be patient and gently as this command may need a few minutes for execution.
}
\examples{
deepblue_find_pattern(pattern = "C[GT]+C",
    genome = "hg19", overlap = FALSE)

}
\seealso{
Other Inserting and listing annotations: \code{\link{deepblue_list_annotations}}
}
