% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalFunctions.R
\name{deepblue_format_object_size}
\alias{deepblue_format_object_size}
\title{Format byte size as human readable units}
\source{
utils:::format.object_size
}
\usage{
deepblue_format_object_size(x, units = "b")
}
\arguments{
\item{x}{size in bytes}

\item{units}{target unit or 'auto'}
}
\value{
formatted size
}
\description{
Format byte size as human readable units
}
