% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue_get_biosource_related}
\alias{deepblue_get_biosource_related}
\title{get_biosource_related}
\usage{
deepblue_get_biosource_related(biosource = NULL,
  user_key = deepblue_USER_KEY)
}
\arguments{
\item{biosource}{- A string (name of the biosource)}

\item{user_key}{- A string (users token key)}
}
\value{
biosources - A array (related biosources)
}
\description{
Gets biosources related to the given one. e.g. the children terms and theirs synonyms.
}
\examples{
deepblue_get_biosource_related(biosource = "Blood")

}
\seealso{
Other Set the relationship between different biosources: \code{\link{deepblue_get_biosource_children}},
  \code{\link{deepblue_get_biosource_parents}},
  \code{\link{deepblue_get_biosource_synonyms}}
}

