% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue_get_biosource_synonyms}
\alias{deepblue_get_biosource_synonyms}
\title{get_biosource_synonyms}
\usage{
deepblue_get_biosource_synonyms(biosource = NULL,
  user_key = deepblue_options("user_key"))
}
\arguments{
\item{biosource}{- A string (name of the biosource)}

\item{user_key}{- A string (users token key)}
}
\value{
synonyms - A array (synonyms of the biosource)
}
\description{
Obtain the synonyms of the specified biosource. Synonyms are different aliases for the same biosource. A BioSource refers to a term describing the origin of a given sample, such as a tissue or cell line.
}
\examples{
deepblue_get_biosource_synonyms(biosource = "prostate gland")

}
\seealso{
Other Set the relationship between different biosources: \code{\link{deepblue_get_biosource_children}},
  \code{\link{deepblue_get_biosource_parents}},
  \code{\link{deepblue_get_biosource_related}}
}

