% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching.R
\name{deepblue_get_db}
\alias{deepblue_get_db}
\title{Sets up the DeepBlueR cache and returns a filehash db object}
\usage{
deepblue_get_db()
}
\value{
A filehash package database
}
\description{
Sets up the DeepBlueR cache and returns a filehash db object
}

