% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue_get_experiments_by_query}
\alias{deepblue_get_experiments_by_query}
\title{get_experiments_by_query}
\usage{
deepblue_get_experiments_by_query(query_id = NULL,
  user_key = deepblue_options("user_key"))
}
\arguments{
\item{query_id}{- A string (id of the query)}

\item{user_key}{- A string (users token key)}
}
\value{
experiments - A array (List containing experiments names and ids)
}
\description{
List the experiments and annotations that have at least one genomic region in the final query result.
}
\examples{
deepblue_get_experiments_by_query(query_id = "q12345")

}
\seealso{
Other Operating on the data regions: \code{\link{deepblue_aggregate}},
  \code{\link{deepblue_binning}},
  \code{\link{deepblue_count_regions}},
  \code{\link{deepblue_coverage}},
  \code{\link{deepblue_extend}},
  \code{\link{deepblue_filter_regions}},
  \code{\link{deepblue_flank}},
  \code{\link{deepblue_get_regions}},
  \code{\link{deepblue_input_regions}},
  \code{\link{deepblue_intersection}},
  \code{\link{deepblue_merge_queries}},
  \code{\link{deepblue_overlap}},
  \code{\link{deepblue_query_cache}},
  \code{\link{deepblue_query_experiment_type}},
  \code{\link{deepblue_score_matrix}},
  \code{\link{deepblue_select_annotations}},
  \code{\link{deepblue_select_experiments}},
  \code{\link{deepblue_select_regions}},
  \code{\link{deepblue_tiling_regions}}
}

