% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue_get_request_data}
\alias{deepblue_get_request_data}
\title{get_request_data}
\usage{
deepblue_get_request_data(request_id = NULL,
  user_key = deepblue_options("user_key"))
}
\arguments{
\item{request_id}{- A string (ID of the request)}

\item{user_key}{- A string (users token key)}
}
\value{
data - A string or a vector of string (the request data)
}
\description{
Download the requested data. The output can be (i) a string (get_regions, score_matrix, and count_regions), or (ii) a list of ID and names (get_experiments_by_query), or (iii) a struct (coverage).
}
\examples{
data_id = deepblue_select_experiments(
    experiment_name="E002-H3K9ac.narrowPeak.bed",
    chromosome="chr1")
request_id = deepblue_get_regions(
    query_id =data_id,
    output_format = "CHROMOSOME,START,END")
deepblue_get_request_data(request_id = request_id)

}
\seealso{
Other Requests status information and results: \code{\link{deepblue_list_requests}}
}

