% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue_input_regions}
\alias{deepblue_input_regions}
\title{input_regions}
\usage{
deepblue_input_regions(genome = NULL, region_set = NULL,
  user_key = deepblue_options("user_key"))
}
\arguments{
\item{genome}{- A string (the target genome)}

\item{region_set}{- A string (Regions in CHROMOSOME    START    END format)}

\item{user_key}{- A string (users token key)}
}
\value{
id - A string (query id)
}
\description{
Upload a set of genomic regions that can be accessed through a query ID. An interesting use case for this command is to upload a set of custom regions for intersecting with genomic regions in DeepBlue to specifically select regions of interest.
}
\examples{
regions_set = "chr1	28735	29810
chr1	135124	135563
chr1	327790	328229
chr1	437151	438164
chr1	449273	450544
chr1	533219	534114
chr1	544738	546649
chr1	713984	714547
chr1	762416	763445
chr1	788863	789211"
deepblue_input_regions(genome = "hg19",
    region_set = regions_set)

}
\seealso{
Other Operating on the data regions: \code{\link{deepblue_aggregate}},
  \code{\link{deepblue_binning}},
  \code{\link{deepblue_count_regions}},
  \code{\link{deepblue_coverage}},
  \code{\link{deepblue_extend}},
  \code{\link{deepblue_filter_regions}},
  \code{\link{deepblue_flank}},
  \code{\link{deepblue_get_experiments_by_query}},
  \code{\link{deepblue_get_regions}},
  \code{\link{deepblue_intersection}},
  \code{\link{deepblue_merge_queries}},
  \code{\link{deepblue_overlap}},
  \code{\link{deepblue_query_cache}},
  \code{\link{deepblue_query_experiment_type}},
  \code{\link{deepblue_score_matrix}},
  \code{\link{deepblue_select_annotations}},
  \code{\link{deepblue_select_experiments}},
  \code{\link{deepblue_select_regions}},
  \code{\link{deepblue_tiling_regions}}
}

