% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue_is_biosource}
\alias{deepblue_is_biosource}
\title{is_biosource}
\usage{
deepblue_is_biosource(biosource_name = NULL, user_key = deepblue_USER_KEY)
}
\arguments{
\item{biosource_name}{- A string (Name of the biosource)}

\item{user_key}{- A string (users token key)}
}
\value{
information - A string or a vector of string (A string containing the biosource name)
}
\description{
Return information for the given biosource name.
}
\examples{
deepblue_is_biosource(biosource_name = "blood")

}
\seealso{
Other Commands for all types of data: \code{\link{deepblue_cancel_request}},
  \code{\link{deepblue_info}},
  \code{\link{deepblue_list_in_use}},
  \code{\link{deepblue_search}}
}

