% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liftover.R
\name{deepblue_liftover}
\alias{deepblue_liftover}
\title{Lift over region results between Genome Assemblies used in DeepBlue}
\usage{
deepblue_liftover(regions, source = "hg19", target = "hg38",
  collapse = TRUE)
}
\arguments{
\item{regions}{The GRanges object to lift over to another assembly}

\item{source}{The source assembly version, e.g. hg38. If NULL, we try to
read the genome version from the GRanges object.}

\item{target}{The target assembly version, e.g. hg19. Required.}

\item{collapse}{Whether to return a single GRanges object or a list of
GRanges (one per region in the input). The latter is the default behavior of
liftOver since multiple assignments are possible.}
}
\value{
A GRanges object using the target chromosome positions
}
\description{
This is a wrapper function for the liftOver function found in
the rtracklayer package. For common genome assemblies available in DeepBlue,
this function automatically downloads the necessary chain file and calls
liftOver.
}
\examples{
data_id = deepblue_select_experiments(
experiment_name="E002-H3K9ac.narrowPeak.bed", chromosome="chr1")
request_id = deepblue_get_regions(query_id =data_id,
                                 output_format = "CHROMOSOME,START,END")
request_data = deepblue_download_request_data(request_id)
deepblue_liftover(request_data, source = "hg38", target = "hg19")
}
