% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue_list_annotations}
\alias{deepblue_list_annotations}
\title{list_annotations}
\usage{
deepblue_list_annotations(genome = NULL, user_key = deepblue_USER_KEY)
}
\arguments{
\item{genome}{- A string or a vector of string (the target genome)}

\item{user_key}{- A string (users token key)}
}
\value{
annotations - A array (annotation ids)
}
\description{
Lists the DeepBlue Annotations.
}
\examples{
deepblue_list_annotations(genome = "hg19")

}

