% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue_list_biosources}
\alias{deepblue_list_biosources}
\title{list_biosources}
\usage{
deepblue_list_biosources(extra_metadata = NULL,
  user_key = deepblue_USER_KEY)
}
\arguments{
\item{extra_metadata}{- A struct (Key-value that must match the biosource extra_metadata.)}

\item{user_key}{- A string (users token key)}
}
\value{
biosources - A array (biosources)
}
\description{
Lists all existing biosources.
}
\examples{
deepblue_list_biosources(extra_metadata = list(ontology_id = "UBERON:0002485"))

}
\seealso{
Other Inserting and listing biosources: \code{\link{deepblue_list_similar_biosources}}
}

