% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue_list_column_types}
\alias{deepblue_list_column_types}
\title{list_column_types}
\usage{
deepblue_list_column_types(user_key = deepblue_USER_KEY)
}
\arguments{
\item{user_key}{- A string (users token key)}
}
\value{
column_types - A array (column types)
}
\description{
Lists all available column types.
}
\examples{
deepblue_list_column_types()

}

