% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue_list_expressions}
\alias{deepblue_list_expressions}
\title{list_expressions}
\usage{
deepblue_list_expressions(expression_type = NULL, sample_id = NULL,
  replica = NULL, project = NULL, user_key = deepblue_USER_KEY)
}
\arguments{
\item{expression_type}{- A string (expression type (supported: 'gene'))}

\item{sample_id}{- A string or a vector of string (sample ID(s))}

\item{replica}{- A int or a vector of int (replica(s))}

\item{project}{- A string or a vector of string (project(s) name)}

\item{user_key}{- A string (users token key)}
}
\value{
expressions - A array (expressions names and IDS)
}
\description{
List the Expression currently available in DeepBlue. A expression is a set of data with an identifier and an expression value.
}
\examples{
deepblue_list_expressions(expression_type='gene')

}
\seealso{
Other Expression data: \code{\link{deepblue_select_expressions}}
}

