% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue_list_gene_expressions}
\alias{deepblue_list_gene_expressions}
\title{list_gene_expressions}
\usage{
deepblue_list_gene_expressions(sample_id = NULL, replica = NULL,
  project = NULL, user_key = deepblue_USER_KEY)
}
\arguments{
\item{sample_id}{- A string or a vector of string (sample ID(s))}

\item{replica}{- A int or a vector of int (replica(s))}

\item{project}{- A string or a vector of string (project(s) name)}

\item{user_key}{- A string (users token key)}
}
\value{
gene_expressions - A array (gene expressions names and IDS)
}
\description{
List all the Gene Expression currently available in DeepBlue. A gene expression is a set of genes with their expression values.
}
\examples{
deepblue_list_gene_expressions()

}
\seealso{
Other Gene models and genes identifiers: \code{\link{deepblue_list_gene_models}},
  \code{\link{deepblue_list_genes}},
  \code{\link{deepblue_select_gene_expressions}},
  \code{\link{deepblue_select_genes}}
}

