% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue_list_gene_models}
\alias{deepblue_list_gene_models}
\title{list_gene_models}
\usage{
deepblue_list_gene_models(user_key = deepblue_options("user_key"))
}
\arguments{
\item{user_key}{- A string (users token key)}
}
\value{
gene_models - A array (gene models names and IDS)
}
\description{
List all the Gene Models currently available in DeepBlue. A gene model is a set of genes usually imported from GENCODE. For example Gencode v22.
}
\examples{
deepblue_list_gene_models()

}
\seealso{
Other Gene models and genes identifiers: \code{\link{deepblue_count_gene_ontology_terms}},
  \code{\link{deepblue_list_genes}},
  \code{\link{deepblue_select_genes}}
}
