% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue_list_genomes}
\alias{deepblue_list_genomes}
\title{list_genomes}
\usage{
deepblue_list_genomes(user_key = deepblue_USER_KEY)
}
\arguments{
\item{user_key}{- A string (users token key)}
}
\value{
genomes - A array (genome names)
}
\description{
Lists all existing genomes.
}
\examples{
deepblue_list_genomes()

}
\seealso{
Other Inserting and listing genomes: \code{\link{deepblue_chromosomes}},
  \code{\link{deepblue_list_similar_genomes}}
}

