% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue_list_projects}
\alias{deepblue_list_projects}
\title{list_projects}
\usage{
deepblue_list_projects(user_key = deepblue_options("user_key"))
}
\arguments{
\item{user_key}{- A string (users token key)}
}
\value{
projects - A array (project names)
}
\description{
List Projects included in DeepBlue.
}
\examples{
deepblue_list_projects()

}
\seealso{
Other Inserting and listing projects: \code{\link{deepblue_list_similar_projects}}
}

