% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue_list_recent_experiments}
\alias{deepblue_list_recent_experiments}
\title{list_recent_experiments}
\usage{
deepblue_list_recent_experiments(days = NULL, genome = NULL,
  epigenetic_mark = NULL, sample = NULL, technique = NULL,
  project = NULL, user_key = deepblue_options("user_key"))
}
\arguments{
\item{days}{- A double (maximum days ago the experiments were added)}

\item{genome}{- A string or a vector of string (the target genome)}

\item{epigenetic_mark}{- A string or a vector of string (name(s) of selected epigenetic mark(s))}

\item{sample}{- A string or a vector of string (id(s) of selected sample(s))}

\item{technique}{- A string or a vector of string (name(s) of selected technique(es))}

\item{project}{- A string or a vector of string (name(s) of selected projects)}

\item{user_key}{- A string (users token key)}
}
\value{
experiments - A array (names of recent experiments)
}
\description{
List the latest Experiments included in DeepBlue that match criteria defined in the parameters. The returned experiments are sorted by insertion date.
}
\examples{
deepblue_list_recent_experiments(days = 2, genome = "hg19")

}
\seealso{
Other Inserting and listing experiments: \code{\link{deepblue_collection_experiments_count}},
  \code{\link{deepblue_faceting_experiments}},
  \code{\link{deepblue_list_experiments}},
  \code{\link{deepblue_list_similar_experiments}},
  \code{\link{deepblue_preview_experiment}}
}

