% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue_list_samples}
\alias{deepblue_list_samples}
\title{list_samples}
\usage{
deepblue_list_samples(biosource = NULL, extra_metadata = NULL,
  user_key = deepblue_options("user_key"))
}
\arguments{
\item{biosource}{- A string or a vector of string (name(s) of selected biosource(s))}

\item{extra_metadata}{- A struct (Metadata that must be matched)}

\item{user_key}{- A string (users token key)}
}
\value{
samples - A array (samples id with their content)
}
\description{
List Samples included in DeepBlue. It is possible to filter by the BioSource and by extra_metadata fields content.
}
\examples{
deepblue_list_samples(biosource = "Blood")

}
