% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue_list_similar_biosources}
\alias{deepblue_list_similar_biosources}
\title{list_similar_biosources}
\usage{
deepblue_list_similar_biosources(name = NULL, user_key = deepblue_USER_KEY)
}
\arguments{
\item{name}{- A string (biosource name)}

\item{user_key}{- A string (users token key)}
}
\value{
biosources - A array (similar biosources)
}
\description{
Lists all BioSources that have a similar name compared to the provided name.
}
\examples{
deepblue_list_similar_biosources(name = "blood")

}
\seealso{
Other Inserting and listing biosources: \code{\link{deepblue_list_biosources}}
}

