% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue_list_similar_experiments}
\alias{deepblue_list_similar_experiments}
\title{list_similar_experiments}
\usage{
deepblue_list_similar_experiments(name = NULL, genome = NULL,
  user_key = deepblue_options("user_key"))
}
\arguments{
\item{name}{- A string (experiment name)}

\item{genome}{- A string or a vector of string (the target genome)}

\item{user_key}{- A string (users token key)}
}
\value{
experiments - A array (similar experiment names)
}
\description{
List all Experiments that have a similar name compared to the provided name. The similarity is calculated using the Levenshtein method.
}
\examples{
deepblue_list_similar_experiments(name = "blood", genome = "hg19")

}
\seealso{
Other Inserting and listing experiments: \code{\link{deepblue_collection_experiments_count}},
  \code{\link{deepblue_faceting_experiments}},
  \code{\link{deepblue_list_experiments}},
  \code{\link{deepblue_list_recent_experiments}},
  \code{\link{deepblue_preview_experiment}}
}

