% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue_list_similar_projects}
\alias{deepblue_list_similar_projects}
\title{list_similar_projects}
\usage{
deepblue_list_similar_projects(name = NULL,
  user_key = deepblue_options("user_key"))
}
\arguments{
\item{name}{- A string (project name)}

\item{user_key}{- A string (users token key)}
}
\value{
projects - A array (similar project names)
}
\description{
List Projects that have a similar name compared to the provided name. The similarity is calculated using the Levenshtein method.
}
\examples{
deepblue_list_similar_projects(name = "BLUEPRINT")

}
\seealso{
Other Inserting and listing projects: \code{\link{deepblue_list_projects}}
}

