% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue_list_similar_techniques}
\alias{deepblue_list_similar_techniques}
\title{list_similar_techniques}
\usage{
deepblue_list_similar_techniques(name = NULL, user_key = deepblue_USER_KEY)
}
\arguments{
\item{name}{- A string (technique name)}

\item{user_key}{- A string (users token key)}
}
\value{
techniques - A array (similar techniques)
}
\description{
Lists all techniques similar to the one provided.
}
\examples{
deepblue_list_similar_techniques(name = "chip seq")

}
\seealso{
Other Inserting and listing techniques: \code{\link{deepblue_list_techniques}}
}

