% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{deepblue_meta_data_to_table}
\alias{deepblue_meta_data_to_table}
\title{Convert XML structured meta data to table format}
\usage{
deepblue_meta_data_to_table(ids, user_key = deepblue_options("user_key"))
}
\arguments{
\item{ids}{an id or a list of ids}

\item{user_key}{a DeepBlue user key (optional for public data)}
}
\value{
a data frame with meta data
}
\description{
Convert XML structured meta data to table format
}
\examples{
#works for sample ids
deepblue_meta_data_to_table(list("s2694", "s2695"))

#or experiment ids
deepblue_meta_data_to_table(list("e30035", "e30036"))
}
