% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue_name_to_id}
\alias{deepblue_name_to_id}
\title{name_to_id}
\usage{
deepblue_name_to_id(name = NULL, collection = NULL,
  user_key = deepblue_options("user_key"))
}
\arguments{
\item{name}{- A string or a vector of string (ID or an array of IDs)}

\item{collection}{- A string (Collection where the data name is in )}

\item{user_key}{- A string (users token key)}
}
\value{
information - A array or a vector of array (List of IDs.)
}
\description{
Obtain the data ID(s) from the informed data name(s).
}
\examples{
deepblue_name_to_id("E002-H3K9ac.narrowPeak.bed", "experiments")
deepblue_name_to_id("prostate duct", "biosources")
deepblue_name_to_id("DNA Methylation", "Epigenetic_marks")

}
\seealso{
Other Commands for all types of data: \code{\link{deepblue_cancel_request}},
  \code{\link{deepblue_info}},
  \code{\link{deepblue_is_biosource}},
  \code{\link{deepblue_list_in_use}},
  \code{\link{deepblue_search}}
}

