% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{deepblue_options}
\alias{deepblue_options}
\title{options}
\usage{
deepblue_options(..., .__defaults = FALSE, .__reset = FALSE)
}
\value{
default options
}
\description{
options manager from the settings package
}

