% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue_overlap}
\alias{deepblue_overlap}
\title{overlap}
\usage{
deepblue_overlap(query_data_id = NULL, query_filter_id = NULL,
  overlap = NULL, amount = NULL, amount_type = NULL,
  user_key = deepblue_options("user_key"))
}
\arguments{
\item{query_data_id}{- A string (query data that will be filtered.)}

\item{query_filter_id}{- A string (query containing the regions that the regions of the query_data_id must overlap.)}

\item{overlap}{- A boolean (True if must overlap, or false if must not overlap.)}

\item{amount}{- A int (Amount of regions that must overlap. Use the parameter 'amount_type' ('bp' or '\%') to specify the unit.  For example, use the value '10' with the amount_type '\%' to specify that 10\% of the bases in both regions must overlap, or use '10' with the amount_type 'bp' to specify that at least 10 bases must or must not overlap.)}

\item{amount_type}{- A string (Type of the amount: 'bp' for base pairs and '\%' for percentage. )}

\item{user_key}{- A string (users token key)}
}
\value{
id - A string (id of the new query)
}
\description{
Select genomic regions that overlap or not overlap with with the specified number of regions of the second query. Important: This command is still experimental and changes may occour.
}
\examples{
annotation_id = deepblue_select_annotations(
    annotation_name="CpG Islands",
    genome="hg19", chromosome="chr1")
experiment_id = deepblue_select_experiments(
    experiment_name="S00XDKH1.ERX712765.H3K27ac.bwa.GRCh38.20150527.bed")
deepblue_overlap(query_data_id = experiment_id, query_filter_id = annotation_id, 
    overlap = TRUE, amount=10, amount_type="\%")


}
\seealso{
Other Operating on the data regions: \code{\link{deepblue_aggregate}},
  \code{\link{deepblue_binning}},
  \code{\link{deepblue_calculate_enrichment}},
  \code{\link{deepblue_count_regions}},
  \code{\link{deepblue_coverage}},
  \code{\link{deepblue_extend}},
  \code{\link{deepblue_filter_regions}},
  \code{\link{deepblue_flank}},
  \code{\link{deepblue_get_experiments_by_query}},
  \code{\link{deepblue_get_regions}},
  \code{\link{deepblue_input_regions}},
  \code{\link{deepblue_intersection}},
  \code{\link{deepblue_merge_queries}},
  \code{\link{deepblue_query_cache}},
  \code{\link{deepblue_query_experiment_type}},
  \code{\link{deepblue_score_matrix}},
  \code{\link{deepblue_select_annotations}},
  \code{\link{deepblue_select_experiments}},
  \code{\link{deepblue_select_regions}},
  \code{\link{deepblue_tiling_regions}}
}
