% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalFunctions.R
\name{deepblue_parse_gtf}
\alias{deepblue_parse_gtf}
\title{deepblue_parse_gtf}
\usage{
deepblue_parse_gtf(all_gtf)
}
\arguments{
\item{all_gtf}{a character vector of GTF attributes for a single region.}
}
\value{
the parsed GTF data
}
\description{
Parse the GTF semicolon separated attributes into a data frame
}
\keyword{internal}
