% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue_query_experiment_type}
\alias{deepblue_query_experiment_type}
\title{query_experiment_type}
\usage{
deepblue_query_experiment_type(query_id = NULL, type = NULL,
  user_key = deepblue_options("user_key"))
}
\arguments{
\item{query_id}{- A string (Query ID)}

\item{type}{- A string (experiment type (peaks or signal))}

\item{user_key}{- A string (users token key)}
}
\value{
information - A string (New query ID.)
}
\description{
Filter the query ID for regions associated with experiments of a given type. For example, it is possible to select only peaks using this command with the 'peaks' parameter.
}
\examples{
h3k27ac_regions = deepblue_select_regions(
    genome ='GRCh38',
    epigenetic_mark ='H3k27ac',
    project ='BLUEPRINT Epigenome',
    chromosome ='chr1')
deepblue_query_experiment_type(
    query_id = h3k27ac_regions,
    type = "peaks")

}
\seealso{
Other Operating on the data regions: \code{\link{deepblue_aggregate}},
  \code{\link{deepblue_binning}},
  \code{\link{deepblue_calculate_enrichment}},
  \code{\link{deepblue_count_regions}},
  \code{\link{deepblue_coverage}},
  \code{\link{deepblue_extend}},
  \code{\link{deepblue_filter_regions}},
  \code{\link{deepblue_flank}},
  \code{\link{deepblue_get_experiments_by_query}},
  \code{\link{deepblue_get_regions}},
  \code{\link{deepblue_input_regions}},
  \code{\link{deepblue_intersection}},
  \code{\link{deepblue_merge_queries}},
  \code{\link{deepblue_overlap}},
  \code{\link{deepblue_query_cache}},
  \code{\link{deepblue_score_matrix}},
  \code{\link{deepblue_select_annotations}},
  \code{\link{deepblue_select_experiments}},
  \code{\link{deepblue_select_regions}},
  \code{\link{deepblue_tiling_regions}}
}
