% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{deepblue_reset_options}
\alias{deepblue_reset_options}
\title{Reset DeepBlueR options}
\usage{
deepblue_reset_options(new_options = NULL)
}
\arguments{
\item{new_options}{list of new options that should be used.
default options if NULL}
}
\value{
new (default) options
}
\description{
Reset DeepBlueR options
}
\examples{
deepblue_reset_options()
}

