% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue_score_matrix}
\alias{deepblue_score_matrix}
\title{score_matrix}
\usage{
deepblue_score_matrix(experiments_columns = NULL,
  aggregation_function = NULL, aggregation_regions_id = NULL,
  user_key = deepblue_options("user_key"))
}
\arguments{
\item{experiments_columns}{- A struct (map with experiments names and columns to be processed. Example : {'wgEncodeBroadHistoneDnd41H3k27acSig.wig':'VALUE', 'wgEncodeBroadHistoneCd20ro01794H3k27acSig.wig':'VALUE'})}

\item{aggregation_function}{- A string (aggregation function name: min, max, sum, mean, var, sd, median, count, boolean)}

\item{aggregation_regions_id}{- A string (query ID of the regions that will be used as the aggregation boundaries)}

\item{user_key}{- A string (users token key)}
}
\value{
score_matrix - A string (the score matrix containing the summarized data)
}
\description{
Build a matrix containing the aggregation result of the the experiments data by the aggregation boundaries.
}
\examples{
tiling_regions = deepblue_tiling_regions(
    size=100000, genome="mm10", chromosome="chr1")
deepblue_score_matrix(
    experiments_columns =
        list(ENCFF721EKA="VALUE", ENCFF781VVH="VALUE"),
  aggregation_function = "mean",
  aggregation_regions_id = tiling_regions)

}
\seealso{
Other Operating on the data regions: \code{\link{deepblue_aggregate}},
  \code{\link{deepblue_binning}},
  \code{\link{deepblue_calculate_enrichment}},
  \code{\link{deepblue_count_regions}},
  \code{\link{deepblue_coverage}},
  \code{\link{deepblue_extend}},
  \code{\link{deepblue_filter_regions}},
  \code{\link{deepblue_flank}},
  \code{\link{deepblue_get_experiments_by_query}},
  \code{\link{deepblue_get_regions}},
  \code{\link{deepblue_input_regions}},
  \code{\link{deepblue_intersection}},
  \code{\link{deepblue_merge_queries}},
  \code{\link{deepblue_overlap}},
  \code{\link{deepblue_query_cache}},
  \code{\link{deepblue_query_experiment_type}},
  \code{\link{deepblue_select_annotations}},
  \code{\link{deepblue_select_experiments}},
  \code{\link{deepblue_select_regions}},
  \code{\link{deepblue_tiling_regions}}
}
