% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{deepblue_select_column}
\alias{deepblue_select_column}
\title{select column}
\usage{
deepblue_select_column(experiments, column,
  user_key = deepblue_options("user_key"))
}
\arguments{
\item{experiments}{- A data frame with experiments obtained from
deepblue_list_experiments}

\item{column}{- The name of the column that is extracted from each experiment
file}

\item{user_key}{- A string (users token key)}
}
\value{
A list of experiments with the selected column
}
\description{
A utility command that creates a list of experiments
in which a specific column is selected. Such a list is needed as input
for deepblue_score_matrix.
}
\examples{
blueprint_DNA_meth <- deepblue_list_experiments(
genome = "GRCh38",
epigenetic_mark = "DNA Methylation",
technique = "Bisulfite-Seq",
project = "BLUEPRINT EPIGENOME")

blueprint_DNA_meth <- blueprint_DNA_meth[grep("bs_call",
 deepblue_extract_names(blueprint_DNA_meth)),]

exp_columns <- deepblue_select_column(blueprint_DNA_meth, "VALUE")
}
\seealso{
\code{\link{deepblue_score_matrix}}

\code{\link{deepblue_list_experiments}}

Other Utilities for information processing: \code{\link{deepblue_diff}}
}
