% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue_select_experiments}
\alias{deepblue_select_experiments}
\title{select_experiments}
\usage{
deepblue_select_experiments(experiment_name = NULL, chromosome = NULL,
  start = NULL, end = NULL, user_key = deepblue_options("user_key"))
}
\arguments{
\item{experiment_name}{- A string or a vector of string (name(s) of selected experiment(s))}

\item{chromosome}{- A string or a vector of string (chromosome name(s))}

\item{start}{- A int (minimum start region)}

\item{end}{- A int (maximum end region)}

\item{user_key}{- A string (users token key)}
}
\value{
id - A string (query id)
}
\description{
Selects regions from Experiments by the experiments names.
}
\examples{
deepblue_select_experiments(
	experiment_name = c("E002-H3K9ac.narrowPeak.bed",
						"E001-H3K4me3.gappedPeak.bed")
)

}
\seealso{
Other Operating on the data regions: \code{\link{deepblue_aggregate}},
  \code{\link{deepblue_binning}},
  \code{\link{deepblue_calculate_enrichment}},
  \code{\link{deepblue_count_regions}},
  \code{\link{deepblue_coverage}},
  \code{\link{deepblue_extend}},
  \code{\link{deepblue_filter_regions}},
  \code{\link{deepblue_flank}},
  \code{\link{deepblue_get_experiments_by_query}},
  \code{\link{deepblue_get_regions}},
  \code{\link{deepblue_input_regions}},
  \code{\link{deepblue_intersection}},
  \code{\link{deepblue_merge_queries}},
  \code{\link{deepblue_overlap}},
  \code{\link{deepblue_query_cache}},
  \code{\link{deepblue_query_experiment_type}},
  \code{\link{deepblue_score_matrix}},
  \code{\link{deepblue_select_annotations}},
  \code{\link{deepblue_select_regions}},
  \code{\link{deepblue_tiling_regions}}
}
