% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue_select_gene_expressions}
\alias{deepblue_select_gene_expressions}
\title{select_gene_expressions}
\usage{
deepblue_select_gene_expressions(sample_ids = NULL, replicas = NULL,
  genes = NULL, projects = NULL, gene_model = NULL,
  user_key = deepblue_USER_KEY)
}
\arguments{
\item{sample_ids}{- A string or a vector of string (id(s) of selected sample(s))}

\item{replicas}{- A int or a vector of int (replica(s))}

\item{genes}{- A string or a vector of string (genes(s) - ensembl ID or ENSB name. It does accept regular expressions. For example: '.*' to select all genes.)}

\item{projects}{- A string or a vector of string (projects(s))}

\item{gene_model}{- A string (gene model name)}

\item{user_key}{- A string (users token key)}
}
\value{
id - A string (query id)
}
\description{
Select genes (by their name or ID) as genomic regions from the specified gene model.
}
\examples{
genes_names =
    c('CCR1', 'CD164', 'CD1D', 'CD2', 'CD34', 'CD3G', 'CD44')
deepblue_select_gene_expressions(
	sample_ids="s10205",
    genes = genes_names,
    gene_model = "gencode v23")

}
\seealso{
Other Gene models and genes identifiers: \code{\link{deepblue_list_gene_expressions}},
  \code{\link{deepblue_list_gene_models}},
  \code{\link{deepblue_list_genes}},
  \code{\link{deepblue_select_genes}}
}

