% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue_select_regions}
\alias{deepblue_select_regions}
\title{select_regions}
\usage{
deepblue_select_regions(experiment_name = NULL, genome = NULL,
  epigenetic_mark = NULL, sample_id = NULL, technique = NULL,
  project = NULL, chromosomes = NULL, start = NULL, end = NULL,
  user_key = deepblue_options("user_key"))
}
\arguments{
\item{experiment_name}{- A string or a vector of string (name(s) of selected experiment(s))}

\item{genome}{- A string or a vector of string (the target genome)}

\item{epigenetic_mark}{- A string or a vector of string (name(s) of selected epigenetic mark(s))}

\item{sample_id}{- A string or a vector of string (id(s) of selected sample(s))}

\item{technique}{- A string or a vector of string (name(s) of selected technique(es))}

\item{project}{- A string or a vector of string (name(s) of selected projects)}

\item{chromosomes}{- A string or a vector of string (chromosome name(s))}

\item{start}{- A int (minimum start region)}

\item{end}{- A int (maximum end region)}

\item{user_key}{- A string (users token key)}
}
\value{
id - A string (query id)
}
\description{
Selects Experiment regions that matches the criteria informed by the operation parameters.
}
\examples{
deepblue_select_regions(
    genome="hg19",
    epigenetic_mark = "H3K27ac",
    project = " BLUEPRINT Epigenome")

}
\seealso{
Other Operating on the data regions: \code{\link{deepblue_aggregate}},
  \code{\link{deepblue_binning}},
  \code{\link{deepblue_calculate_enrichment}},
  \code{\link{deepblue_count_regions}},
  \code{\link{deepblue_coverage}},
  \code{\link{deepblue_extend}},
  \code{\link{deepblue_filter_regions}},
  \code{\link{deepblue_flank}},
  \code{\link{deepblue_get_experiments_by_query}},
  \code{\link{deepblue_get_regions}},
  \code{\link{deepblue_input_regions}},
  \code{\link{deepblue_intersection}},
  \code{\link{deepblue_merge_queries}},
  \code{\link{deepblue_overlap}},
  \code{\link{deepblue_query_cache}},
  \code{\link{deepblue_query_experiment_type}},
  \code{\link{deepblue_score_matrix}},
  \code{\link{deepblue_select_annotations}},
  \code{\link{deepblue_select_experiments}},
  \code{\link{deepblue_tiling_regions}}
}
