% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue_tiling_regions}
\alias{deepblue_tiling_regions}
\title{tiling_regions}
\usage{
deepblue_tiling_regions(size = NULL, genome = NULL, chromosome = NULL,
  user_key = deepblue_options("user_key"))
}
\arguments{
\item{size}{- A int (tiling size)}

\item{genome}{- A string (the target genome)}

\item{chromosome}{- A string or a vector of string (chromosome name(s))}

\item{user_key}{- A string (users token key)}
}
\value{
id - A string (query id)
}
\description{
Generate tiling regions across the genome chromosomes. The idea is to "bin" genomic regions systematically in order to obtain discrete regions over which one can aggregate. Using the 'score_matrix' command, these bins (tiles) can be compared directly across experiments.
}
\examples{
deepblue_tiling_regions(
    size = 10000,
    genome = "hg19",
    chromosome = "chr1")

}
\seealso{
Other Operating on the data regions: \code{\link{deepblue_aggregate}},
  \code{\link{deepblue_binning}},
  \code{\link{deepblue_calculate_enrichment}},
  \code{\link{deepblue_count_regions}},
  \code{\link{deepblue_coverage}},
  \code{\link{deepblue_extend}},
  \code{\link{deepblue_filter_regions}},
  \code{\link{deepblue_flank}},
  \code{\link{deepblue_get_experiments_by_query}},
  \code{\link{deepblue_get_regions}},
  \code{\link{deepblue_input_regions}},
  \code{\link{deepblue_intersection}},
  \code{\link{deepblue_merge_queries}},
  \code{\link{deepblue_overlap}},
  \code{\link{deepblue_query_cache}},
  \code{\link{deepblue_query_experiment_type}},
  \code{\link{deepblue_score_matrix}},
  \code{\link{deepblue_select_annotations}},
  \code{\link{deepblue_select_experiments}},
  \code{\link{deepblue_select_regions}}
}
