% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\docType{data}
\name{deepblue_USER_KEY}
\alias{deepblue_USER_KEY}
\title{default User Key}
\format{An object of class \code{character} of length 1.}
\usage{
deepblue_USER_KEY
}
\value{
default user key for accessing DeepBlue
}
\description{
Default anonymous user key
}
\keyword{internal}

