% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_DeepBlueCommand.R
\docType{class}
\name{DeepBlueCommand-class}
\alias{DeepBlueCommand-class}
\alias{DeepBlueCommand}
\title{DeepBlueCommand class}
\arguments{
\item{call}{language}

\item{status}{character}

\item{query_id}{character}

\item{previous_commands}{list}

\item{user_key}{character}

\item{result}{GRanges}
}
\value{
class for managin DeepBlue commands
}
\description{
An S4 class returned when calling
a DeepBlue-R function. It holds information about
the original call, the query / request status,
previous commands, the user_key, and results in GRanges format once
a request is downloaded.
}
