% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpingFunctions.R
\name{get_request_data}
\alias{get_request_data}
\title{get_request_data}
\usage{
get_request_data(request_info, user_key = deepblue.USER_KEY,
  type = "grange")
}
\arguments{
\item{request_info}{A list}

\item{user}{A string}
}
\value{
grange_regions Final output in GRanges format
}
\description{
Returns the requested data as GRanges object. Expects two input parameters; Request information and
user key. It depends on outputs from several functions, namely;
deepblue.get_request_data, convert_to_df, and convert_to_grange.
}
\seealso{
\code{\link{deepblue.get_request_data}}, \code{\link{convert_to_df}}, and
\code{\link{convert_to_grange}}.
}

