% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpingFunctions.R
\name{process_request}
\alias{process_request}
\title{process_request}
\usage{
process_request(request_id, sleep.time = 1, user_key = deepblue.USER_KEY)
}
\arguments{
\item{request_id}{A string with the request_id}

\item{sleep.time}{An integer with default value 1s}

\item{user_key}{A string}
}
\description{
Process the user request. Takes in three parameters; requested regions, sleep time, and user key.
}

