% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_DeepBlueCommand.R
\docType{methods}
\name{show,DeepBlueCommand-method}
\alias{show,DeepBlueCommand-method}
\title{show DeepBlue command object}
\usage{
\S4method{show}{DeepBlueCommand}(object)
}
\arguments{
\item{object}{DeepBlueCommand object}
}
\value{
query ID of the object
}
\description{
Provides a summary of a DeepBlue command object
}
\keyword{internal}
