% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{xml.rpc}
\alias{xml.rpc}
\title{xml.rpc}
\usage{
xml.rpc(url, method, ..., .args = list(...), .opts = list(),
  .defaultOpts = list(httpheader = c(`Content-Type` = "text/xml"),
  followlocation = TRUE, useragent = useragent), .convert = TRUE,
  .curl = getCurlHandle(), useragent = "DeepBlue-R-XMLRPC",
  verbose = deepblue_options("debug"))
}
\value{
XML RPC request data converted to R objects
}
\description{
perform an XML-RPC call
}
\keyword{internal}

