% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{xmlRPCToR.array}
\alias{xmlRPCToR.array}
\title{xmlRPCToR.struct}
\usage{
xmlRPCToR.array(node, ...)
}
\arguments{
\item{node}{- XML-RPC node that contains the array.}
}
\description{
convert XML-RPC array to R data'
}

