% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{xmlRPCToR.struct}
\alias{xmlRPCToR.struct}
\title{xmlRPCToR.struct}
\usage{
xmlRPCToR.struct(node, ...)
}
\arguments{
\item{node}{- XML-RPC node that contains the struct.}
}
\description{
convert XML-RPC struct to R data'
}

