% Generated by roxygen2: do not edit by hand
% Please edit documentation in inst/deepblue.R
\name{deepblue.add_annotation}
\alias{deepblue.add_annotation}
\title{add_annotation}
\usage{
deepblue.add_annotation(name = NULL, genome = NULL, description = NULL,
  data = NULL, format = NULL, extra_metadata = NULL,
  user_key = deepblue.USER_KEY)
}
\arguments{
\item{name}{- A string (annotation name)}

\item{genome}{- A string (the target genome)}

\item{description}{- A string (description of the annotation)}

\item{data}{- A string (the BED formatted data)}

\item{format}{- A string (format of the provided data)}

\item{extra_metadata}{- A struct (additional metadata)}

\item{user_key}{- A string (users token key)}
}
\value{
id - A string (id of the newly inserted annotation)
}
\description{
Inserts a new annotation with the given parameters.
}
\seealso{
Other Inserting and listing annotations: \code{\link{deepblue.find_pattern}},
  \code{\link{deepblue.list_annotations}}
}

