% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue.add_biosource}
\alias{deepblue.add_biosource}
\title{add_biosource}
\usage{
deepblue.add_biosource(name = NULL, description = NULL,
  extra_metadata = NULL, user_key = deepblue.USER_KEY)
}
\arguments{
\item{name}{- A string (biosource name)}

\item{description}{- A string (description of the biosource)}

\item{extra_metadata}{- A struct (additional metadata)}

\item{user_key}{- A string (users token key)}
}
\value{
id - A string (id of the newly inserted biosource)
}
\description{
Inserts a new biosource with the given parameters.
}
\examples{
deepblue.add_biosource(name="New BioSource",
	description="This Biosource is used for..",
	extra_metadata=list(source="Ontology Blabla",
		reference="http://ontology.blabla/id"),
	user_key="my_private_user_key")

}
\seealso{
Other Inserting and listing biosources: \code{\link{deepblue.list_biosources}},
  \code{\link{deepblue.list_similar_biosources}}
}

