% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue.add_genome}
\alias{deepblue.add_genome}
\title{add_genome}
\usage{
deepblue.add_genome(name = NULL, description = NULL, data = NULL,
  user_key = deepblue.USER_KEY)
}
\arguments{
\item{name}{- A string (genome name)}

\item{description}{- A string (description of the genome)}

\item{data}{- A string (genome data)}

\item{user_key}{- A string (users token key)}
}
\value{
id - A string (id of the newly inserted genome)
}
\description{
Inserts a new genome with the given parameters.
}
\examples{
\dontrun{
  GRCh38_info = "chr1	248956422
chr2	242193529
chr3	198295559
chr4	190214555
chr5	181538259
chr6	170805979
chr7	159345973
chr8	145138636
chr9	138394717
chr10	133797422
chr11	135086622
chr12	133275309
chr13	114364328
chr14	107043718
chr15	101991189
chr16	90338345
chr17	83257441
chr18	80373285
chr19	58617616
chr20	64444167
chr21	46709983
chr22	50818468
chrX	156040895
chrY	57227415
chrM	16569"

deepblue.add_genome(name = "GRCh38",
	description = "GRCh38 with the main chromosomes",
	data = GRCh38_info, user_key = "my_secret_key")
}

}
\seealso{
Other Inserting and listing genomes: \code{\link{deepblue.chromosomes}},
  \code{\link{deepblue.list_genomes}},
  \code{\link{deepblue.list_similar_genomes}},
  \code{\link{deepblue.upload_chromosome}}
}

