% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue.add_sample}
\alias{deepblue.add_sample}
\title{add_sample}
\usage{
deepblue.add_sample(biosource_name = NULL, extra_metadata = NULL,
  user_key = deepblue.USER_KEY)
}
\arguments{
\item{biosource_name}{- A string (biosource name)}

\item{extra_metadata}{- A struct (sample extra metadata. You can include any key-value collection here.)}

\item{user_key}{- A string (users token key)}
}
\value{
id - A string (id of the newly inserted sample)
}
\description{
Inserts a new sample of a given biosourcea.
}
\examples{
\dontrun{
  deepblue.add_sample(biosource_name = "BLOOD",
	extra_metadata = list(age="20 days"), user_key = "my_secret_key")
}

}
\seealso{
Other Inserting and listing samples: \code{\link{deepblue.add_sample_from_gsm}},
  \code{\link{deepblue.list_samples}}
}

