% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue.add_technique}
\alias{deepblue.add_technique}
\title{add_technique}
\usage{
deepblue.add_technique(name = NULL, description = NULL,
  extra_metadata = NULL, user_key = deepblue.USER_KEY)
}
\arguments{
\item{name}{- A string (technique name)}

\item{description}{- A string (description of technique)}

\item{extra_metadata}{- A struct (additional metadata)}

\item{user_key}{- A string (users token key)}
}
\value{
id - A string (id of the newly inserted technique)
}
\description{
Inserts a technique with the given parameters.
}
\examples{
\dontrun{
  deepblue.add_technique(name = "New technique",
	description = "This technique is faster and cheaper",
	extra_metadata = list(information = "some useful information"),
	user_key = "my_secret_key")
}

}
\seealso{
Other Inserting and listing techniques: \code{\link{deepblue.list_similar_techniques}},
  \code{\link{deepblue.list_techniques}}
}

