% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchExport.R
\name{deepblue.batch_export_results}
\alias{deepblue.batch_export_results}
\title{batch_export_results}
\usage{
deepblue.batch_export_results(requests, target.directory = NULL,
  suffix = "", prefix = "", user_key = deepblue.USER_KEY)
}
\arguments{
\item{requests}{A list of request objects}

\item{target.directory}{Where the results should be saved}

\item{user_key}{A string used to authenticate the user}
}
\description{
Write results from DeepBlue to disk as they become available
}

