% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.r
\name{deepblue.commands}
\alias{deepblue.commands}
\title{commands}
\usage{
deepblue.commands()
}
\value{
a vector with  ('okay',
}
\description{
Lists all existing commands.

command descriptions) or ('error', error_message)
}
\seealso{
Other Checking DeepBlue status: \code{\link{deepblue.echo}}
}

