% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalFunctions.R
\name{deepblue.convert_to_df}
\alias{deepblue.convert_to_df}
\title{convert_to_df}
\usage{
deepblue.convert_to_df(string_to_parse, request_info, dict = col_dict)
}
\arguments{
\item{string_to_parse}{A string}

\item{request_info}{The request information returned by DeepBlue}

\item{dict}{The data structure that contains the DeepBlue columns types}
}
\value{
regions A data frame
}
\description{
save output in a data frame for further processing.
Expects two parameters; the output string from method
deepblue.get_request_data and request information from method process_request.
}
\keyword{internal}

