% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.r
\name{deepblue.create_column_type_simple}
\alias{deepblue.create_column_type_simple}
\title{create_column_type_simple}
\usage{
deepblue.create_column_type_simple(name = NULL, description = NULL,
  type = NULL, user_key = deepblue.USER_KEY)
}
\arguments{
\item{name}{- A string (column type name)}

\item{description}{- A string (description of the column type)}

\item{type}{- A string (type of the column type)}

\item{user_key}{- A string (users token key)}
}
\value{
a vector with  ('okay',
}
\description{
Create a column type from a category set.

id of the newly created column type) or ('error', error_message)
}
\seealso{
Other Inserting and listing different column types: \code{\link{deepblue.create_column_type_calculated}},
  \code{\link{deepblue.create_column_type_category}},
  \code{\link{deepblue.create_column_type_range}},
  \code{\link{deepblue.list_column_types}}
}

