% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalFunctions.R
\name{deepblue.download_request_data}
\alias{deepblue.download_request_data}
\title{deepblue.download_request_data}
\usage{
deepblue.download_request_data(request_id, user_key = deepblue.USER_KEY,
  type = "grange")
}
\arguments{
\item{request_id}{- Id of the request that will be downloaded}

\item{user_key}{A string}

\item{type}{To which type the downloaded data will be converted:
'string', 'df', or 'grange'.}
}
\value{
grange_regions Final output in GRanges format
}
\description{
Returns the requested data as the expected type object.
Expects two input parameters; Request information and
user key. It depends on outputs from several functions, namely;
deepblue.get_request_data, convert_to_df, and convert_to_grange.
}
\examples{
data_id = deepblue.select_experiments(
experiment_name="E002-H3K9ac.narrowPeak.bed", chromosome="chr1")
request_id = deepblue.get_regions(query_id =data_id,
  output_format = "CHROMOSOME,START,END")
request_data = deepblue.download_request_data(request_id)
}

