% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue.list_projects}
\alias{deepblue.list_projects}
\title{list_projects}
\usage{
deepblue.list_projects(user_key = deepblue.USER_KEY)
}
\arguments{
\item{user_key}{- A string (users token key)}
}
\value{
projects - A array (project names)
}
\description{
Lists all existing projects.
}
\examples{
\dontrun{
  deepblue.list_projects()
}

}
\seealso{
Other Inserting and listing projects: \code{\link{deepblue.add_project}},
  \code{\link{deepblue.add_user_to_project}},
  \code{\link{deepblue.list_similar_projects}},
  \code{\link{deepblue.set_project_public}}
}

