% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue.list_requests}
\alias{deepblue.list_requests}
\title{list_requests}
\usage{
deepblue.list_requests(request_state = NULL, user_key = deepblue.USER_KEY)
}
\arguments{
\item{request_state}{- A string (Name of the state to get requests for. The valid states are: new, running, done, and failed.)}

\item{user_key}{- A string (users token key)}
}
\value{
data_state - A array (Request-IDs and their state)
}
\description{
Lists all requests in given state.
}
\examples{
deepblue.list_requests(request_state = 'done')

}
\seealso{
Other Requests status information and results: \code{\link{deepblue.get_request_data}}
}

