% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue.list_samples}
\alias{deepblue.list_samples}
\title{list_samples}
\usage{
deepblue.list_samples(biosource = NULL, extra_metadata = NULL,
  user_key = deepblue.USER_KEY)
}
\arguments{
\item{biosource}{- A string or a vector of string (biosource name)}

\item{extra_metadata}{- A struct (Key-value that must match the sample extra_metadata.)}

\item{user_key}{- A string (users token key)}
}
\value{
samples - A array (samples id with their content)
}
\description{
Lists all existing samples that matches the given biosource and metadata.
}
\examples{
deepblue.list_samples(biosource = "Blood")

}

