% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue.query_cache}
\alias{deepblue.query_cache}
\title{query_cache}
\usage{
deepblue.query_cache(query_id = NULL, cache = NULL,
  user_key = deepblue.USER_KEY)
}
\arguments{
\item{query_id}{- A string (query ID)}

\item{cache}{- A boolean (set or unset this query caching)}

\item{user_key}{- A string (users token key)}
}
\value{
information - A string (New query ID.)
}
\description{
Return information for the given ID (or IDs).
}
\examples{
annotation_id = deepblue.select_annotations(
    annotation_name="CpG Islands",
    genome="hg19", chromosome="chr1")
data_id = deepblue.select_experiments(
    experiment_name="E002-H3K9ac.narrowPeak.bed")
merged_regions = deepblue.merge_queries(
    query_a_id = annotation_id,
    query_b_id = data_id)
deepblue.query_cache(
    query_id = merged_regions, cache = TRUE)

}
\seealso{
Other Operating on the data regions: \code{\link{deepblue.aggregate}},
  \code{\link{deepblue.count_regions}},
  \code{\link{deepblue.extend}},
  \code{\link{deepblue.filter_regions}},
  \code{\link{deepblue.flank}},
  \code{\link{deepblue.get_experiments_by_query}},
  \code{\link{deepblue.get_regions}},
  \code{\link{deepblue.input_regions}},
  \code{\link{deepblue.intersection}},
  \code{\link{deepblue.merge_queries}},
  \code{\link{deepblue.query_experiment_type}},
  \code{\link{deepblue.score_matrix}},
  \code{\link{deepblue.select_annotations}},
  \code{\link{deepblue.select_experiments}},
  \code{\link{deepblue.select_regions}},
  \code{\link{deepblue.tiling_regions}}
}

