% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue.remove}
\alias{deepblue.remove}
\title{remove}
\usage{
deepblue.remove(id = NULL, user_key = deepblue.USER_KEY)
}
\arguments{
\item{id}{- A string (Data ID to be removed.)}

\item{user_key}{- A string (users token key)}
}
\value{
id - A string (id of the removed data)
}
\description{
Remove data from DeepBlue.
}
\examples{
\dontrun{
  deepblue.remove(id = data_id, user_key = "my_secret_key")
}

}
\seealso{
Other Commands for all types of data: \code{\link{deepblue.cancel_request}},
  \code{\link{deepblue.info}},
  \code{\link{deepblue.is_biosource}},
  \code{\link{deepblue.list_in_use}},
  \code{\link{deepblue.search}}
}

