% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue.set_biosource_synonym}
\alias{deepblue.set_biosource_synonym}
\title{set_biosource_synonym}
\usage{
deepblue.set_biosource_synonym(biosource = NULL, synonym_name = NULL,
  user_key = deepblue.USER_KEY)
}
\arguments{
\item{biosource}{- A string (name of the biosource)}

\item{synonym_name}{- A string (name of the synonym)}

\item{user_key}{- A string (users token key)}
}
\value{
synonym_name - A string (inserted synonym_name)
}
\description{
Sets a biosource synonym.
}
\examples{
\dontrun{
  deepblue.set_biosource_synonym(biosource = "biosource name",
  synonym_name = "synonym name", user_key = "my_secret_key")
}

}
\seealso{
Other Set the relationship between different biosources: \code{\link{deepblue.get_biosource_children}},
  \code{\link{deepblue.get_biosource_parents}},
  \code{\link{deepblue.get_biosource_related}},
  \code{\link{deepblue.get_biosource_synonyms}},
  \code{\link{deepblue.set_biosource_parent}}
}

