% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue_count_regions}
\alias{deepblue_count_regions}
\title{count_regions}
\usage{
deepblue_count_regions(query_id = NULL,
  user_key = deepblue_options("user_key"))
}
\arguments{
\item{query_id}{- A string (Query ID)}

\item{user_key}{- A string (users token key)}
}
\value{
request_id - A string (Request ID - Use it to retrieve the result with info() and get_request_data())
}
\description{
Return the number of genomic regions present in the query.
}
\examples{
data_id = deepblue_select_experiments(
    experiment_name="E002-H3K9ac.narrowPeak.bed")
deepblue_count_regions(query_id = data_id)

}
\seealso{
Other Operating on the data regions: \code{\link{deepblue_aggregate}},
  \code{\link{deepblue_binning}},
  \code{\link{deepblue_coverage}},
  \code{\link{deepblue_extend}},
  \code{\link{deepblue_filter_regions}},
  \code{\link{deepblue_flank}},
  \code{\link{deepblue_get_experiments_by_query}},
  \code{\link{deepblue_get_regions}},
  \code{\link{deepblue_input_regions}},
  \code{\link{deepblue_intersection}},
  \code{\link{deepblue_merge_queries}},
  \code{\link{deepblue_overlap}},
  \code{\link{deepblue_query_cache}},
  \code{\link{deepblue_query_experiment_type}},
  \code{\link{deepblue_score_matrix}},
  \code{\link{deepblue_select_annotations}},
  \code{\link{deepblue_select_experiments}},
  \code{\link{deepblue_select_regions}},
  \code{\link{deepblue_tiling_regions}}
}

