% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue_echo}
\alias{deepblue_echo}
\title{echo}
\usage{
deepblue_echo(user_key = deepblue_options("user_key"))
}
\arguments{
\item{user_key}{- A string (users token key)}
}
\value{
message - A string (echo message including version)
}
\description{
Greet the user with the DeepBlue version.
}
\examples{
deepblue_echo(user_key = "anonymous_key")

}
\seealso{
Other Checking DeepBlue status: \code{\link{deepblue_commands}}
}

