% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{deepblue_extract_ids}
\alias{deepblue_extract_ids}
\title{extract_ids}
\usage{
deepblue_extract_ids(df = NULL)
}
\arguments{
\item{df}{- A array of IDs and names}
}
\value{
ids - A vector containing the extracted IDs)
}
\description{
A utility command that returns a list of IDs extracted
from a data frame of ID and names.
}
\examples{
deepblue_extract_ids(
    df = data.frame(id = c("a124", "a1235"),
    name = c("Annotation 1", "Annotation 2")))

}
\seealso{
Other Utilities for connecting operations: \code{\link{deepblue_extract_names}}
}

