% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{deepblue_extract_names}
\alias{deepblue_extract_names}
\title{extract_names}
\usage{
deepblue_extract_names(df = NULL)
}
\arguments{
\item{df}{- A array of IDs and Names}
}
\value{
names - A vector containing the extracted names
}
\description{
A utility command that returns a list of names extracted from a list of ID and names.
}
\examples{
deepblue_extract_ids(
    df = data.frame(id = c("a124", "a1235"),
    name = c("Annotation 1", "Annotation 2")))

}
\seealso{
Other Utilities for connecting operations: \code{\link{deepblue_extract_ids}}
}
