% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue_get_biosource_children}
\alias{deepblue_get_biosource_children}
\title{get_biosource_children}
\usage{
deepblue_get_biosource_children(biosource = NULL,
  user_key = deepblue_options("user_key"))
}
\arguments{
\item{biosource}{- A string (biosource name)}

\item{user_key}{- A string (users token key)}
}
\value{
biosources - A array (related biosources)
}
\description{
A BioSource refers to a term describing the origin of a given sample, such as a tissue or cell line. These form a hierarchy in which children of a BioSource term can be fetched with this command. Children terms are more specific terms that are defined in the imported ontologies.
}
\examples{
deepblue_get_biosource_children(biosource = "Blood")

}
\seealso{
Other Set the relationship between different biosources: \code{\link{deepblue_get_biosource_parents}},
  \code{\link{deepblue_get_biosource_related}},
  \code{\link{deepblue_get_biosource_synonyms}}
}

