% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue_get_biosource_parents}
\alias{deepblue_get_biosource_parents}
\title{get_biosource_parents}
\usage{
deepblue_get_biosource_parents(biosource = NULL,
  user_key = deepblue_options("user_key"))
}
\arguments{
\item{biosource}{- A string (biosource name)}

\item{user_key}{- A string (users token key)}
}
\value{
biosources - A array (parents biosources)
}
\description{
A BioSource refers to a term describing the origin of a given sample, such as a tissue or cell line. These form a hierarchy in which the parent of a BioSource term can be fetched with this command. Parent terms are more generic terms that are defined in the imported ontologies.
}
\examples{
deepblue_get_biosource_parents(biosource = "Blood")

}
\seealso{
Other Set the relationship between different biosources: \code{\link{deepblue_get_biosource_children}},
  \code{\link{deepblue_get_biosource_related}},
  \code{\link{deepblue_get_biosource_synonyms}}
}

