% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue_get_biosource_related}
\alias{deepblue_get_biosource_related}
\title{get_biosource_related}
\usage{
deepblue_get_biosource_related(biosource = NULL,
  user_key = deepblue_options("user_key"))
}
\arguments{
\item{biosource}{- A string (biosource name)}

\item{user_key}{- A string (users token key)}
}
\value{
biosources - A array (related biosources)
}
\description{
A BioSource refers to a term describing the origin of a given sample, such as a tissue or cell line. These form a hierarchy in which the children of a BioSource term and its synonyms can be fetched with this command. Children terms are more specific terms that are defined in the imported ontologies. Synonyms are different aliases for the same biosource.
}
\examples{
deepblue_get_biosource_related(biosource = "Blood")

}
\seealso{
Other Set the relationship between different biosources: \code{\link{deepblue_get_biosource_children}},
  \code{\link{deepblue_get_biosource_parents}},
  \code{\link{deepblue_get_biosource_synonyms}}
}
