% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue_is_biosource}
\alias{deepblue_is_biosource}
\title{is_biosource}
\usage{
deepblue_is_biosource(biosource = NULL,
  user_key = deepblue_options("user_key"))
}
\arguments{
\item{biosource}{- A string (biosource name)}

\item{user_key}{- A string (users token key)}
}
\value{
information - A string or a vector of string (A string containing the biosource name)
}
\description{
Verify if the name is an existing and valid DeepBlue BioSource name. A BioSource refers to a term describing the origin of a given sample, such as a tissue or cell line.
}
\examples{
deepblue_is_biosource(biosource = "blood")

}
\seealso{
Other Commands for all types of data: \code{\link{deepblue_cancel_request}},
  \code{\link{deepblue_info}},
  \code{\link{deepblue_list_in_use}},
  \code{\link{deepblue_name_to_id}},
  \code{\link{deepblue_search}}
}
