% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue_list_epigenetic_marks}
\alias{deepblue_list_epigenetic_marks}
\title{list_epigenetic_marks}
\usage{
deepblue_list_epigenetic_marks(extra_metadata = NULL,
  user_key = deepblue_options("user_key"))
}
\arguments{
\item{extra_metadata}{- A struct (Metadata that must be matched)}

\item{user_key}{- A string (users token key)}
}
\value{
epigenetic_marks - A array (epigenetic mark names and IDS)
}
\description{
List Epigenetic Marks included in DeepBlue. This includes histone marks, DNA methylation, DNA sensitivity, etc. It is possible to filter the Epigenetic Marks by their extra_metadata field content.
}
\examples{
deepblue_list_epigenetic_marks()

}
\seealso{
Other Inserting and listing epigenetic marks: \code{\link{deepblue_list_similar_epigenetic_marks}}
}

