% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue_list_genes}
\alias{deepblue_list_genes}
\title{list_genes}
\usage{
deepblue_list_genes(gene_models = NULL, user_key = deepblue_USER_KEY)
}
\arguments{
\item{gene_models}{- A string or a vector of string (the gene model)}

\item{user_key}{- A string (users token key)}
}
\value{
genes - A array (genes names and IDs)
}
\description{
List all the Gene currently available in DeepBlue.
}
\examples{
deepblue.list_genes('Gencode v22')

}
\seealso{
Other Operations on gene sets and genes identifiers: \code{\link{deepblue_add_gene_model}},
  \code{\link{deepblue_list_gene_models}},
  \code{\link{deepblue_select_genes}}
}

