% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue_list_genes}
\alias{deepblue_list_genes}
\title{list_genes}
\usage{
deepblue_list_genes(genes = NULL, chromosome = NULL, start = NULL,
  end = NULL, gene_models = NULL, user_key = deepblue_options("user_key"))
}
\arguments{
\item{genes}{- A string or a vector of string (Name(s) or ENSEMBL ID (ENSGXXXXXXXXXXX.X ) of the gene(s).)}

\item{chromosome}{- A string or a vector of string (chromosome name(s))}

\item{start}{- A int (minimum start region)}

\item{end}{- A int (maximum end region)}

\item{gene_models}{- A string or a vector of string (the gene model)}

\item{user_key}{- A string (users token key)}
}
\value{
genes - A array (genes names and its content)
}
\description{
List the Genes currently available in DeepBlue.
}
\examples{
deepblue_list_genes(
  chromosome="chr20", 
  start=10000000, 
  end=21696620, 
gene_models='Gencode v22')

}
\seealso{
Other Gene models and genes identifiers: \code{\link{deepblue_list_gene_models}},
  \code{\link{deepblue_select_genes}}
}

