% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue_list_in_use}
\alias{deepblue_list_in_use}
\title{list_in_use}
\usage{
deepblue_list_in_use(controlled_vocabulary = NULL,
  user_key = deepblue_options("user_key"))
}
\arguments{
\item{controlled_vocabulary}{- A string (controlled vocabulary name)}

\item{user_key}{- A string (users token key)}
}
\value{
terms - A array (controlled_vocabulary terms with count)
}
\description{
List all terms used by the Experiments mandatory metadata that have at least one Experiment or Annotation using them.
}
\examples{
deepblue_list_in_use(controlled_vocabulary = "biosources")

}
\seealso{
Other Commands for all types of data: \code{\link{deepblue_cancel_request}},
  \code{\link{deepblue_info}},
  \code{\link{deepblue_is_biosource}},
  \code{\link{deepblue_name_to_id}},
  \code{\link{deepblue_search}}
}

