% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue_list_similar_biosources}
\alias{deepblue_list_similar_biosources}
\title{list_similar_biosources}
\usage{
deepblue_list_similar_biosources(name = NULL,
  user_key = deepblue_options("user_key"))
}
\arguments{
\item{name}{- A string (biosource name)}

\item{user_key}{- A string (users token key)}
}
\value{
biosource - A string (biosource name)
}
\description{
List all BioSources that have a similar name compared to the provided name. A BioSource refers to a term describing the origin of a given sample, such as a tissue or cell line. The similarity is calculated using the Levenshtein method.
}
\examples{
deepblue_list_similar_biosources(name = "blood")

}
\seealso{
Other Inserting and listing biosources: \code{\link{deepblue_list_biosources}}
}
