% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue_list_similar_genomes}
\alias{deepblue_list_similar_genomes}
\title{list_similar_genomes}
\usage{
deepblue_list_similar_genomes(name = NULL,
  user_key = deepblue_options("user_key"))
}
\arguments{
\item{name}{- A string (genome name)}

\item{user_key}{- A string (users token key)}
}
\value{
genomes - A array (similar genome names)
}
\description{
Lists all Genomes that have a similar name compared to the provided name. The similarity is calculated using the Levenshtein method.
}
\examples{
deepblue_list_similar_genomes(name = "grc")

}
\seealso{
Other Inserting and listing genomes: \code{\link{deepblue_chromosomes}},
  \code{\link{deepblue_list_genomes}}
}

