% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue_list_similar_techniques}
\alias{deepblue_list_similar_techniques}
\title{list_similar_techniques}
\usage{
deepblue_list_similar_techniques(name = NULL,
  user_key = deepblue_options("user_key"))
}
\arguments{
\item{name}{- A string (technique name)}

\item{user_key}{- A string (users token key)}
}
\value{
techniques - A array (similar techniques)
}
\description{
List Techniques that have a similar name compared to the provided name. The similarity is calculated using the Levenshtein method.
}
\examples{
deepblue_list_similar_techniques(name = "chip seq")

}
\seealso{
Other Inserting and listing techniques: \code{\link{deepblue_list_techniques}}
}

