% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue_list_techniques}
\alias{deepblue_list_techniques}
\title{list_techniques}
\usage{
deepblue_list_techniques(user_key = deepblue_options("user_key"))
}
\arguments{
\item{user_key}{- A string (users token key)}
}
\value{
techniques - A array (techniques)
}
\description{
List the Techniques included in DeepBlue.
}
\examples{
deepblue_list_techniques()

}
\seealso{
Other Inserting and listing techniques: \code{\link{deepblue_list_similar_techniques}}
}

