% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.R
\name{deepblue_select_genes}
\alias{deepblue_select_genes}
\title{select_genes}
\usage{
deepblue_select_genes(genes = NULL, gene_model = NULL, chromosome = NULL,
  start = NULL, end = NULL, user_key = deepblue_USER_KEY)
}
\arguments{
\item{genes}{- A string or a vector of string (Name(s) or ENSEMBL ID (ENSGXXXXXXXXXXX.X ) of the gene(s).)}

\item{gene_model}{- A string (gene model name)}

\item{chromosome}{- A string or a vector of string (chromosome name(s))}

\item{start}{- A int (minimum start region)}

\item{end}{- A int (maximum end region)}

\item{user_key}{- A string (users token key)}
}
\value{
id - A string (query id)
}
\description{
Select genes (by their name or ID) as genomic regions from the specified gene model.
}
\examples{
genes_names =
    c('CCR1', 'CD164', 'CD1D', 'CD2', 'CD34', 'CD3G', 'CD44')
deepblue_select_genes(
    genes = genes_names,
    gene_model = "gencode v23")

}
\seealso{
Other Gene models and genes identifiers: \code{\link{deepblue_list_gene_models}},
  \code{\link{deepblue_list_genes}}
}

